--- pkg/frontend/nvme.go
@@ -1,650 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// Copyright (c) 2022-2023 Dell Inc, or its subsidiaries.
-// Copyright (c) 2022 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-// Copyright (C) 2023 Intel Corporation
-
-// Package frontend implememnts the FrontEnd APIs (host facing) of the storage Server
-package frontend
-
-import (
-	"context"
-	"fmt"
-	"log"
-	"path"
-	"sort"
-	"strconv"
-
-	"github.com/opiproject/gospdk/spdk"
-
-	pb "github.com/opiproject/opi-api/storage/v1alpha1/gen/go"
-	"github.com/opiproject/opi-nvidia-bridge/pkg/models"
-	"github.com/opiproject/opi-spdk-bridge/pkg/server"
-
-	"github.com/google/uuid"
-	"go.einride.tech/aip/fieldmask"
-	"go.einride.tech/aip/resourceid"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
-	"google.golang.org/protobuf/types/known/emptypb"
-)
-
-func sortNvmeSubsystems(subsystems []*pb.NvmeSubsystem) {
-	sort.Slice(subsystems, func(i int, j int) bool {
-		return subsystems[i].Spec.Nqn < subsystems[j].Spec.Nqn
-	})
-}
-
-func sortNvmeControllers(controllers []*pb.NvmeController) {
-	sort.Slice(controllers, func(i int, j int) bool {
-		return controllers[i].Spec.NvmeControllerId < controllers[j].Spec.NvmeControllerId
-	})
-}
-
-func sortNvmeNamespaces(namespaces []*pb.NvmeNamespace) {
-	sort.Slice(namespaces, func(i int, j int) bool {
-		return namespaces[i].Spec.HostNsid < namespaces[j].Spec.HostNsid
-	})
-}
-
-// CreateNvmeSubsystem creates an Nvme Subsystem
-func (s *Server) CreateNvmeSubsystem(_ context.Context, in *pb.CreateNvmeSubsystemRequest) (*pb.NvmeSubsystem, error) {
-	log.Printf("CreateNvmeSubsystem: Received from client: %v", in)
-	// see https://google.aip.dev/133#user-specified-ids
-	resourceID := resourceid.NewSystemGenerated()
-	if in.NvmeSubsystemId != "" {
-		err := resourceid.ValidateUserSettable(in.NvmeSubsystemId)
-		if err != nil {
-			log.Printf("error: %v", err)
-			return nil, err
-		}
-		log.Printf("client provided the ID of a resource %v, ignoring the name field %v", in.NvmeSubsystemId, in.NvmeSubsystem.Name)
-		resourceID = in.NvmeSubsystemId
-	}
-	in.NvmeSubsystem.Name = server.ResourceIDToVolumeName(resourceID)
-	// idempotent API when called with same key, should return same object
-	subsys, ok := s.Subsystems[in.NvmeSubsystem.Name]
-	if ok {
-		log.Printf("Already existing NvmeSubsystem with id %v", in.NvmeSubsystem.Name)
-		return subsys, nil
-	}
-	// check if another object exists with same NQN, it is not allowed
-	for _, item := range s.Subsystems {
-		if in.NvmeSubsystem.Spec.Nqn == item.Spec.Nqn {
-			msg := fmt.Sprintf("Could not create NQN: %s since object %s with same NQN already exists", in.NvmeSubsystem.Spec.Nqn, item.Name)
-			log.Print(msg)
-			return nil, status.Errorf(codes.AlreadyExists, msg)
-		}
-	}
-	// not found, so create a new one
-	params := models.NvdaSubsystemNvmeCreateParams{
-		Nqn:          in.NvmeSubsystem.Spec.Nqn,
-		SerialNumber: in.NvmeSubsystem.Spec.SerialNumber,
-		ModelNumber:  in.NvmeSubsystem.Spec.ModelNumber,
-	}
-	var result models.NvdaSubsystemNvmeCreateResult
-	err := s.rpc.Call("subsystem_nvme_create", &params, &result)
-	if err != nil {
-		log.Printf("error: %v", err)
-		return nil, err
-	}
-	log.Printf("Received from SPDK: %v", result)
-	if !result {
-		msg := fmt.Sprintf("Could not create NQN: %s", in.NvmeSubsystem.Spec.Nqn)
-		log.Print(msg)
-		return nil, status.Errorf(codes.InvalidArgument, msg)
-	}
-	var ver spdk.GetVersionResult
-	err = s.rpc.Call("spdk_get_version", nil, &ver)
-	if err != nil {
-		log.Printf("error: %v", err)
-		return nil, err
-	}
-	log.Printf("Received from SPDK: %v", ver)
-	response := server.ProtoClone(in.NvmeSubsystem)
-	response.Status = &pb.NvmeSubsystemStatus{FirmwareRevision: ver.Version}
-	s.Subsystems[in.NvmeSubsystem.Name] = response
-	return response, nil
-}
-
-// DeleteNvmeSubsystem deletes an Nvme Subsystem
-func (s *Server) DeleteNvmeSubsystem(_ context.Context, in *pb.DeleteNvmeSubsystemRequest) (*emptypb.Empty, error) {
-	log.Printf("DeleteNvmeSubsystem: Received from client: %v", in)
-	subsys, ok := s.Subsystems[in.Name]
-	if !ok {
-		if in.AllowMissing {
-			return &emptypb.Empty{}, nil
-		}
-		err := status.Errorf(codes.NotFound, "unable to find key %s", in.Name)
-		log.Printf("error: %v", err)
-		return nil, err
-	}
-	params := models.NvdaSubsystemNvmeDeleteParams{
-		Nqn: subsys.Spec.Nqn,
-	}
-	var result models.NvdaSubsystemNvmeDeleteResult
-	err := s.rpc.Call("subsystem_nvme_delete", &params, &result)
-	if err != nil {
-		log.Printf("error: %v", err)
-		return nil, err
-	}
-	log.Printf("Received from SPDK: %v", result)
-	if !result {
-		msg := fmt.Sprintf("Could not delete NQN: %s", subsys.Spec.Nqn)
-		log.Print(msg)
-		return nil, status.Errorf(codes.InvalidArgument, msg)
-	}
-	delete(s.Subsystems, subsys.Name)
-	return &emptypb.Empty{}, nil
-}
-
-// UpdateNvmeSubsystem updates an Nvme Subsystem
-func (s *Server) UpdateNvmeSubsystem(_ context.Context, in *pb.UpdateNvmeSubsystemRequest) (*pb.NvmeSubsystem, error) {
-	log.Printf("UpdateNvmeSubsystem: Received from client: %v", in)
-	volume, ok := s.Subsystems[in.NvmeSubsystem.Name]
-	if !ok {
-		if in.AllowMissing {
-			log.Printf("TODO: in case of AllowMissing, create a new resource, don;t return error")
-		}
-		err := status.Errorf(codes.NotFound, "unable to find key %s", in.NvmeSubsystem.Name)
-		log.Printf("error: %v", err)
-		return nil, err
-	}
-	resourceID := path.Base(volume.Name)
-	// update_mask = 2
-	if err := fieldmask.Validate(in.UpdateMask, in.NvmeSubsystem); err != nil {
-		log.Printf("error: %v", err)
-		return nil, err
-	}
-	log.Printf("TODO: use resourceID=%v", resourceID)
-	return nil, status.Errorf(codes.Unimplemented, "UpdateNvmeSubsystem method is not implemented")
-}
-
-// ListNvmeSubsystems lists Nvme Subsystems
-func (s *Server) ListNvmeSubsystems(_ context.Context, in *pb.ListNvmeSubsystemsRequest) (*pb.ListNvmeSubsystemsResponse, error) {
-	log.Printf("ListNvmeSubsystems: Received from client: %v", in)
-	size, offset, perr := server.ExtractPagination(in.PageSize, in.PageToken, s.Pagination)
-	if perr != nil {
-		log.Printf("error: %v", perr)
-		return nil, perr
-	}
-	var result []models.NvdaSubsystemNvmeListResult
-	err := s.rpc.Call("subsystem_nvme_list", nil, &result)
-	if err != nil {
-		log.Printf("error: %v", err)
-		return nil, err
-	}
-	log.Printf("Received from SPDK: %v", result)
-	token, hasMoreElements := "", false
-	log.Printf("Limiting result len(%d) to [%d:%d]", len(result), offset, size)
-	result, hasMoreElements = server.LimitPagination(result, offset, size)
-	if hasMoreElements {
-		token = uuid.New().String()
-		s.Pagination[token] = offset + size
-	}
-	Blobarray := make([]*pb.NvmeSubsystem, len(result))
-	for i := range result {
-		r := &result[i]
-		Blobarray[i] = &pb.NvmeSubsystem{Spec: &pb.NvmeSubsystemSpec{Nqn: r.Nqn, SerialNumber: r.SerialNumber, ModelNumber: r.ModelNumber}}
-	}
-	sortNvmeSubsystems(Blobarray)
-	return &pb.ListNvmeSubsystemsResponse{NvmeSubsystems: Blobarray}, nil
-}
-
-// GetNvmeSubsystem gets Nvme Subsystems
-func (s *Server) GetNvmeSubsystem(_ context.Context, in *pb.GetNvmeSubsystemRequest) (*pb.NvmeSubsystem, error) {
-	log.Printf("GetNvmeSubsystem: Received from client: %v", in)
-	subsys, ok := s.Subsystems[in.Name]
-	if !ok {
-		err := status.Errorf(codes.NotFound, "unable to find key %s", in.Name)
-		log.Printf("error: %v", err)
-		return nil, err
-	}
-	var result []models.NvdaSubsystemNvmeListResult
-	err := s.rpc.Call("subsystem_nvme_list", nil, &result)
-	if err != nil {
-		log.Printf("error: %v", err)
-		return nil, err
-	}
-	log.Printf("Received from SPDK: %v", result)
-	for i := range result {
-		r := &result[i]
-		if r.Nqn == subsys.Spec.Nqn {
-			return &pb.NvmeSubsystem{Spec: &pb.NvmeSubsystemSpec{Nqn: r.Nqn, SerialNumber: r.SerialNumber, ModelNumber: r.ModelNumber}, Status: &pb.NvmeSubsystemStatus{FirmwareRevision: "TBD"}}, nil
-		}
-	}
-	msg := fmt.Sprintf("Could not find NQN: %s", subsys.Spec.Nqn)
-	log.Print(msg)
-	return nil, status.Errorf(codes.InvalidArgument, msg)
-}
-
-// NvmeSubsystemStats gets Nvme Subsystem stats
-func (s *Server) NvmeSubsystemStats(_ context.Context, in *pb.NvmeSubsystemStatsRequest) (*pb.NvmeSubsystemStatsResponse, error) {
-	log.Printf("NvmeSubsystemStats: Received from client: %v", in)
-	return nil, status.Errorf(codes.Unimplemented, "UpdateNvmeSubsystem method is not implemented")
-}
-
-// CreateNvmeController creates an Nvme controller
-func (s *Server) CreateNvmeController(_ context.Context, in *pb.CreateNvmeControllerRequest) (*pb.NvmeController, error) {
-	log.Printf("CreateNvmeController: Received from client: %v", in)
-	// check input parameters validity
-	if in.NvmeController.Spec == nil || in.NvmeController.Spec.SubsystemId == nil || in.NvmeController.Spec.SubsystemId.Value == "" {
-		return nil, status.Error(codes.InvalidArgument, "invalid input subsystem parameters")
-	}
-	// see https://google.aip.dev/133#user-specified-ids
-	resourceID := resourceid.NewSystemGenerated()
-	if in.NvmeControllerId != "" {
-		err := resourceid.ValidateUserSettable(in.NvmeControllerId)
-		if err != nil {
-			log.Printf("error: %v", err)
-			return nil, err
-		}
-		log.Printf("client provided the ID of a resource %v, ignoring the name field %v", in.NvmeControllerId, in.NvmeController.Name)
-		resourceID = in.NvmeControllerId
-	}
-	in.NvmeController.Name = server.ResourceIDToVolumeName(resourceID)
-	// idempotent API when called with same key, should return same object
-	controller, ok := s.Controllers[in.NvmeController.Name]
-	if ok {
-		log.Printf("Already existing NvmeController with id %v", in.NvmeController.Name)
-		return controller, nil
-	}
-	// not found, so create a new one
-	subsys, ok := s.Subsystems[in.NvmeController.Spec.SubsystemId.Value]
-	if !ok {
-		err := status.Errorf(codes.NotFound, "unable to find key %s", in.NvmeController.Spec.SubsystemId.Value)
-		log.Printf("error: %v", err)
-		return nil, err
-	}
-
-	params := models.NvdaControllerNvmeCreateParams{
-		Nqn:              subsys.Spec.Nqn,
-		EmulationManager: "mlx5_0",
-		PfID:             int(in.NvmeController.Spec.PcieId.PhysicalFunction),
-		// VfID:             int(in.NvmeController.Spec.PcieId.VirtualFunction),
-		// MaxNamespaces:    int(in.NvmeController.Spec.MaxNsq),
-		// NrIoQueues:       int(in.NvmeController.Spec.MaxNcq),
-	}
-	var result models.NvdaControllerNvmeCreateResult
-	err := s.rpc.Call("controller_nvme_create", &params, &result)
-	if err != nil {
-		log.Printf("error: %v", err)
-		return nil, err
-	}
-	log.Printf("Received from SPDK: %v", result)
-	if result.Cntlid < 0 {
-		msg := fmt.Sprintf("Could not create CTRL: %s", in.NvmeController.Name)
-		log.Print(msg)
-		return nil, status.Errorf(codes.InvalidArgument, msg)
-	}
-	response := server.ProtoClone(in.NvmeController)
-	response.Spec.NvmeControllerId = int32(result.Cntlid)
-	response.Status = &pb.NvmeControllerStatus{Active: true}
-	s.Controllers[in.NvmeController.Name] = response
-	return response, nil
-}
-
-// DeleteNvmeController deletes an Nvme controller
-func (s *Server) DeleteNvmeController(_ context.Context, in *pb.DeleteNvmeControllerRequest) (*emptypb.Empty, error) {
-	log.Printf("DeleteNvmeController: Received from client: %v", in)
-	controller, ok := s.Controllers[in.Name]
-	if !ok {
-		if in.AllowMissing {
-			return &emptypb.Empty{}, nil
-		}
-		return nil, fmt.Errorf("error finding controller %s", in.Name)
-	}
-	subsys, ok := s.Subsystems[controller.Spec.SubsystemId.Value]
-	if !ok {
-		err := status.Errorf(codes.NotFound, "unable to find key %s", controller.Spec.SubsystemId.Value)
-		log.Printf("error: %v", err)
-		return nil, err
-	}
-
-	params := models.NvdaControllerNvmeDeleteParams{
-		Subnqn: subsys.Spec.Nqn,
-		Cntlid: int(controller.Spec.NvmeControllerId),
-	}
-	var result models.NvdaControllerNvmeDeleteResult
-	err := s.rpc.Call("controller_nvme_delete", &params, &result)
-	if err != nil {
-		log.Printf("error: %v", err)
-		return nil, err
-	}
-	log.Printf("Received from SPDK: %v", result)
-	if !result {
-		msg := fmt.Sprintf("Could not delete NQN:ID %s:%d", subsys.Spec.Nqn, controller.Spec.NvmeControllerId)
-		log.Print(msg)
-		return nil, status.Errorf(codes.InvalidArgument, msg)
-	}
-	delete(s.Controllers, controller.Name)
-	return &emptypb.Empty{}, nil
-}
-
-// UpdateNvmeController updates an Nvme controller
-func (s *Server) UpdateNvmeController(_ context.Context, in *pb.UpdateNvmeControllerRequest) (*pb.NvmeController, error) {
-	log.Printf("UpdateNvmeController: Received from client: %v", in)
-	volume, ok := s.Controllers[in.NvmeController.Name]
-	if !ok {
-		if in.AllowMissing {
-			log.Printf("TODO: in case of AllowMissing, create a new resource, don;t return error")
-		}
-		err := status.Errorf(codes.NotFound, "unable to find key %s", in.NvmeController.Name)
-		log.Printf("error: %v", err)
-		return nil, err
-	}
-	resourceID := path.Base(volume.Name)
-	// update_mask = 2
-	if err := fieldmask.Validate(in.UpdateMask, in.NvmeController); err != nil {
-		log.Printf("error: %v", err)
-		return nil, err
-	}
-	log.Printf("TODO: use resourceID=%v", resourceID)
-	return nil, status.Errorf(codes.Unimplemented, "UpdateNvmeController method is not implemented")
-}
-
-// ListNvmeControllers lists Nvme controllers
-func (s *Server) ListNvmeControllers(_ context.Context, in *pb.ListNvmeControllersRequest) (*pb.ListNvmeControllersResponse, error) {
-	log.Printf("ListNvmeControllers: Received from client: %v", in)
-	size, offset, perr := server.ExtractPagination(in.PageSize, in.PageToken, s.Pagination)
-	if perr != nil {
-		log.Printf("error: %v", perr)
-		return nil, perr
-	}
-	subsys, ok := s.Subsystems[in.Parent]
-	if !ok {
-		err := status.Errorf(codes.NotFound, "unable to find key %s", in.Parent)
-		log.Printf("error: %v", err)
-		return nil, err
-	}
-	var result []models.NvdaControllerListResult
-	err := s.rpc.Call("controller_list", nil, &result)
-	if err != nil {
-		log.Printf("error: %v", err)
-		return nil, err
-	}
-	log.Printf("Received from SPDK: %v", result)
-	token, hasMoreElements := "", false
-	log.Printf("Limiting result len(%d) to [%d:%d]", len(result), offset, size)
-	result, hasMoreElements = server.LimitPagination(result, offset, size)
-	if hasMoreElements {
-		token = uuid.New().String()
-		s.Pagination[token] = offset + size
-	}
-	Blobarray := make([]*pb.NvmeController, len(result))
-	for i := range result {
-		r := &result[i]
-		if r.Subnqn == subsys.Spec.Nqn && r.Type == "nvme" {
-			Blobarray[i] = &pb.NvmeController{Spec: &pb.NvmeControllerSpec{NvmeControllerId: int32(r.Cntlid)}}
-		}
-	}
-	sortNvmeControllers(Blobarray)
-	return &pb.ListNvmeControllersResponse{NvmeControllers: Blobarray}, nil
-}
-
-// GetNvmeController gets an Nvme controller
-func (s *Server) GetNvmeController(_ context.Context, in *pb.GetNvmeControllerRequest) (*pb.NvmeController, error) {
-	log.Printf("GetNvmeController: Received from client: %v", in)
-	controller, ok := s.Controllers[in.Name]
-	if !ok {
-		err := status.Errorf(codes.NotFound, "unable to find key %s", in.Name)
-		log.Printf("error: %v", err)
-		return nil, err
-	}
-	var result []models.NvdaControllerListResult
-	err := s.rpc.Call("controller_list", nil, &result)
-	if err != nil {
-		log.Printf("error: %v", err)
-		return nil, err
-	}
-	log.Printf("Received from SPDK: %v", result)
-	for i := range result {
-		r := &result[i]
-		if r.Cntlid == int(controller.Spec.NvmeControllerId) && r.Type == "nvme" {
-			return &pb.NvmeController{Spec: &pb.NvmeControllerSpec{NvmeControllerId: int32(r.Cntlid)}, Status: &pb.NvmeControllerStatus{Active: true}}, nil
-		}
-	}
-	msg := fmt.Sprintf("Could not find NvmeControllerId: %d", controller.Spec.NvmeControllerId)
-	log.Print(msg)
-	return nil, status.Errorf(codes.InvalidArgument, msg)
-}
-
-// NvmeControllerStats gets an Nvme controller stats
-func (s *Server) NvmeControllerStats(_ context.Context, in *pb.NvmeControllerStatsRequest) (*pb.NvmeControllerStatsResponse, error) {
-	log.Printf("NvmeControllerStats: Received from client: %v", in)
-	return nil, status.Errorf(codes.Unimplemented, "NvmeControllerStats method is not implemented")
-}
-
-// CreateNvmeNamespace creates an Nvme namespace
-func (s *Server) CreateNvmeNamespace(_ context.Context, in *pb.CreateNvmeNamespaceRequest) (*pb.NvmeNamespace, error) {
-	log.Printf("CreateNvmeNamespace: Received from client: %v", in)
-	// check input parameters validity
-	if in.NvmeNamespace.Spec == nil || in.NvmeNamespace.Spec.SubsystemId == nil || in.NvmeNamespace.Spec.SubsystemId.Value == "" {
-		return nil, status.Error(codes.InvalidArgument, "invalid input subsystem parameters")
-	}
-	// see https://google.aip.dev/133#user-specified-ids
-	resourceID := resourceid.NewSystemGenerated()
-	if in.NvmeNamespaceId != "" {
-		err := resourceid.ValidateUserSettable(in.NvmeNamespaceId)
-		if err != nil {
-			log.Printf("error: %v", err)
-			return nil, err
-		}
-		log.Printf("client provided the ID of a resource %v, ignoring the name field %v", in.NvmeNamespaceId, in.NvmeNamespace.Name)
-		resourceID = in.NvmeNamespaceId
-	}
-	in.NvmeNamespace.Name = server.ResourceIDToVolumeName(resourceID)
-	// idempotent API when called with same key, should return same object
-	namespace, ok := s.Namespaces[in.NvmeNamespace.Name]
-	if ok {
-		log.Printf("Already existing NvmeNamespace with id %v", in.NvmeNamespace.Name)
-		return namespace, nil
-	}
-	// not found, so create a new one
-	subsys, ok := s.Subsystems[in.NvmeNamespace.Spec.SubsystemId.Value]
-	if !ok {
-		err := status.Errorf(codes.NotFound, "unable to find key %s", in.NvmeNamespace.Spec.SubsystemId.Value)
-		log.Printf("error: %v", err)
-		return nil, err
-	}
-	// TODO: do lookup through VolumeId key instead of using it's value
-	params := models.NvdaControllerNvmeNamespaceAttachParams{
-		BdevType: "spdk",
-		Bdev:     in.NvmeNamespace.Spec.VolumeId.Value,
-		Nsid:     int(in.NvmeNamespace.Spec.HostNsid),
-		Subnqn:   subsys.Spec.Nqn,
-		Cntlid:   0,
-		UUID:     in.NvmeNamespace.Spec.Uuid.Value,
-		Nguid:    in.NvmeNamespace.Spec.Nguid,
-		Eui64:    strconv.FormatInt(in.NvmeNamespace.Spec.Eui64, 10),
-	}
-	var result models.NvdaControllerNvmeNamespaceAttachResult
-	err := s.rpc.Call("controller_nvme_namespace_attach", &params, &result)
-	if err != nil {
-		log.Printf("error: %v", err)
-		return nil, err
-	}
-	log.Printf("Received from SPDK: %v", result)
-	if !result {
-		msg := fmt.Sprintf("Could not create NS: %s", in.NvmeNamespace.Name)
-		log.Print(msg)
-		return nil, status.Errorf(codes.InvalidArgument, msg)
-	}
-	response := server.ProtoClone(in.NvmeNamespace)
-	response.Status = &pb.NvmeNamespaceStatus{PciState: 2, PciOperState: 1}
-	s.Namespaces[in.NvmeNamespace.Name] = response
-	return response, nil
-}
-
-// DeleteNvmeNamespace deletes an Nvme namespace
-func (s *Server) DeleteNvmeNamespace(_ context.Context, in *pb.DeleteNvmeNamespaceRequest) (*emptypb.Empty, error) {
-	log.Printf("DeleteNvmeNamespace: Received from client: %v", in)
-	namespace, ok := s.Namespaces[in.Name]
-	if !ok {
-		if in.AllowMissing {
-			return &emptypb.Empty{}, nil
-		}
-		err := status.Errorf(codes.NotFound, "unable to find key %s", in.Name)
-		log.Printf("error: %v", err)
-		return nil, err
-	}
-	subsys, ok := s.Subsystems[namespace.Spec.SubsystemId.Value]
-	if !ok {
-		err := fmt.Errorf("unable to find subsystem %s", namespace.Spec.SubsystemId.Value)
-		log.Printf("error: %v", err)
-		return nil, err
-	}
-
-	// TODO: fix hard-coded Cntlid
-	params := models.NvdaControllerNvmeNamespaceDetachParams{
-		Nsid:   int(namespace.Spec.HostNsid),
-		Subnqn: subsys.Spec.Nqn,
-		Cntlid: 0,
-	}
-	var result models.NvdaControllerNvmeNamespaceDetachResult
-	err := s.rpc.Call("controller_nvme_namespace_detach", &params, &result)
-	if err != nil {
-		log.Printf("error: %v", err)
-		return nil, err
-	}
-	log.Printf("Received from SPDK: %v", result)
-	if !result {
-		msg := fmt.Sprintf("Could not delete NS: %s", namespace.Spec.SubsystemId.Value)
-		log.Print(msg)
-		return nil, status.Errorf(codes.InvalidArgument, msg)
-	}
-	delete(s.Namespaces, namespace.Name)
-	return &emptypb.Empty{}, nil
-}
-
-// UpdateNvmeNamespace updates an Nvme namespace
-func (s *Server) UpdateNvmeNamespace(_ context.Context, in *pb.UpdateNvmeNamespaceRequest) (*pb.NvmeNamespace, error) {
-	log.Printf("UpdateNvmeNamespace: Received from client: %v", in)
-	volume, ok := s.Namespaces[in.NvmeNamespace.Name]
-	if !ok {
-		if in.AllowMissing {
-			log.Printf("TODO: in case of AllowMissing, create a new resource, don;t return error")
-		}
-		err := status.Errorf(codes.NotFound, "unable to find key %s", in.NvmeNamespace.Name)
-		log.Printf("error: %v", err)
-		return nil, err
-	}
-	resourceID := path.Base(volume.Name)
-	// update_mask = 2
-	if err := fieldmask.Validate(in.UpdateMask, in.NvmeNamespace); err != nil {
-		log.Printf("error: %v", err)
-		return nil, err
-	}
-	log.Printf("TODO: use resourceID=%v", resourceID)
-	return nil, status.Errorf(codes.Unimplemented, "UpdateNvmeNamespace method is not implemented")
-}
-
-// ListNvmeNamespaces lists Nvme namespaces
-func (s *Server) ListNvmeNamespaces(_ context.Context, in *pb.ListNvmeNamespacesRequest) (*pb.ListNvmeNamespacesResponse, error) {
-	log.Printf("ListNvmeNamespaces: Received from client: %v", in)
-	size, offset, perr := server.ExtractPagination(in.PageSize, in.PageToken, s.Pagination)
-	if perr != nil {
-		log.Printf("error: %v", perr)
-		return nil, perr
-	}
-	subsys, ok := s.Subsystems[in.Parent]
-	if !ok {
-		err := status.Errorf(codes.NotFound, "unable to find key %s", in.Parent)
-		log.Printf("error: %v", err)
-		return nil, err
-	}
-	// TODO: fix hard-coded Cntlid
-	params := models.NvdaControllerNvmeNamespaceListParams{
-		Subnqn: subsys.Spec.Nqn,
-		Cntlid: 0,
-	}
-	var result models.NvdaControllerNvmeNamespaceListResult
-	err := s.rpc.Call("controller_nvme_namespace_list", &params, &result)
-	if err != nil {
-		log.Printf("error: %v", err)
-		return nil, err
-	}
-	log.Printf("Received from SPDK: %v", result)
-	token, hasMoreElements := "", false
-	log.Printf("Limiting result len(%d) to [%d:%d]", len(result.Namespaces), offset, size)
-	result.Namespaces, hasMoreElements = server.LimitPagination(result.Namespaces, offset, size)
-	if hasMoreElements {
-		token = uuid.New().String()
-		s.Pagination[token] = offset + size
-	}
-	Blobarray := make([]*pb.NvmeNamespace, len(result.Namespaces))
-	for i := range result.Namespaces {
-		r := &result.Namespaces[i]
-		Blobarray[i] = &pb.NvmeNamespace{Spec: &pb.NvmeNamespaceSpec{HostNsid: int32(r.Nsid)}}
-	}
-	sortNvmeNamespaces(Blobarray)
-	return &pb.ListNvmeNamespacesResponse{NvmeNamespaces: Blobarray}, nil
-}
-
-// GetNvmeNamespace gets an Nvme namespace
-func (s *Server) GetNvmeNamespace(_ context.Context, in *pb.GetNvmeNamespaceRequest) (*pb.NvmeNamespace, error) {
-	log.Printf("GetNvmeNamespace: Received from client: %v", in)
-	namespace, ok := s.Namespaces[in.Name]
-	if !ok {
-		err := status.Errorf(codes.NotFound, "unable to find key %s", in.Name)
-		log.Printf("error: %v", err)
-		return nil, err
-	}
-	subsys, ok := s.Subsystems[namespace.Spec.SubsystemId.Value]
-	if !ok {
-		err := status.Errorf(codes.NotFound, "unable to find key %s", namespace.Spec.SubsystemId.Value)
-		log.Printf("error: %v", err)
-		return nil, err
-	}
-	// TODO: fix hard-coded Cntlid
-	params := models.NvdaControllerNvmeNamespaceListParams{
-		Subnqn: subsys.Spec.Nqn,
-		Cntlid: 0,
-	}
-	var result models.NvdaControllerNvmeNamespaceListResult
-	err := s.rpc.Call("controller_nvme_namespace_list", &params, &result)
-	if err != nil {
-		log.Printf("error: %v", err)
-		return nil, err
-	}
-	log.Printf("Received from SPDK: %v", result)
-	for i := range result.Namespaces {
-		r := &result.Namespaces[i]
-		if r.Nsid == int(namespace.Spec.HostNsid) {
-			return &pb.NvmeNamespace{Spec: &pb.NvmeNamespaceSpec{HostNsid: int32(r.Nsid)}, Status: &pb.NvmeNamespaceStatus{PciState: 2, PciOperState: 1}}, nil
-		}
-	}
-	msg := fmt.Sprintf("Could not find HostNsid: %d", namespace.Spec.HostNsid)
-	log.Print(msg)
-	return nil, status.Errorf(codes.InvalidArgument, msg)
-}
-
-// NvmeNamespaceStats gets an Nvme namespace stats
-func (s *Server) NvmeNamespaceStats(_ context.Context, in *pb.NvmeNamespaceStatsRequest) (*pb.NvmeNamespaceStatsResponse, error) {
-	log.Printf("NvmeNamespaceStats: Received from client: %v", in)
-	namespace, ok := s.Namespaces[in.NamespaceId.Value]
-	if !ok {
-		err := status.Errorf(codes.NotFound, "unable to find key %s", in.NamespaceId.Value)
-		log.Printf("error: %v", err)
-		return nil, err
-	}
-	var result models.NvdaControllerNvmeStatsResult
-	err := s.rpc.Call("controller_nvme_get_iostat", nil, &result)
-	if err != nil {
-		log.Printf("error: %v", err)
-		return nil, err
-	}
-	log.Printf("Received from SPDK: %v", result)
-	for _, c := range result.Controllers {
-		for _, r := range c.Bdevs {
-			if r.BdevName == namespace.Spec.VolumeId.Value {
-				return &pb.NvmeNamespaceStatsResponse{Id: in.NamespaceId, Stats: &pb.VolumeStats{
-					ReadOpsCount:  int32(r.ReadIos),
-					WriteOpsCount: int32(r.WriteIos),
-				}}, nil
-			}
-		}
-	}
-	msg := fmt.Sprintf("Could not find BdevName: %s", namespace.Spec.VolumeId.Value)
-	log.Print(msg)
-	return nil, status.Errorf(codes.InvalidArgument, msg)
-}
