--- pkg/frontend/nvme_test.go
@@ -1,2178 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// Copyright (c) 2022-2023 Dell Inc, or its subsidiaries.
-// Copyright (c) 2022 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-
-// Package frontend implememnts the FrontEnd APIs (host facing) of the storage Server
-package frontend
-
-import (
-	"bytes"
-	"fmt"
-	"reflect"
-	"testing"
-
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
-	"google.golang.org/protobuf/proto"
-	"google.golang.org/protobuf/types/known/emptypb"
-	"google.golang.org/protobuf/types/known/fieldmaskpb"
-
-	pc "github.com/opiproject/opi-api/common/v1/gen/go"
-	pb "github.com/opiproject/opi-api/storage/v1alpha1/gen/go"
-	"github.com/opiproject/opi-spdk-bridge/pkg/server"
-)
-
-func TestFrontEnd_CreateNvmeSubsystem(t *testing.T) {
-	spec := &pb.NvmeSubsystemSpec{
-		Nqn:          "nqn.2022-09.io.spdk:opi3",
-		SerialNumber: "OpiSerialNumber",
-		ModelNumber:  "OpiModelNumber",
-	}
-	tests := map[string]struct {
-		id      string
-		in      *pb.NvmeSubsystem
-		out     *pb.NvmeSubsystem
-		spdk    []string
-		errCode codes.Code
-		errMsg  string
-		start   bool
-		exist   bool
-	}{
-		"illegal resource_id": {
-			"CapitalLettersNotAllowed",
-			&pb.NvmeSubsystem{
-				Spec: spec,
-			},
-			nil,
-			[]string{""},
-			codes.Unknown,
-			fmt.Sprintf("user-settable ID must only contain lowercase, numbers and hyphens (%v)", "got: 'C' in position 0"),
-			false,
-			false,
-		},
-		"valid request with invalid SPDK response": {
-			testSubsystemID,
-			&pb.NvmeSubsystem{
-				Spec: spec,
-			},
-			nil,
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":false}`},
-			codes.InvalidArgument,
-			fmt.Sprintf("Could not create NQN: %v", "nqn.2022-09.io.spdk:opi3"),
-			true,
-			false,
-		},
-		"valid request with empty SPDK response": {
-			testSubsystemID,
-			&pb.NvmeSubsystem{
-				Spec: spec,
-			},
-			nil,
-			[]string{""},
-			codes.Unknown,
-			fmt.Sprintf("subsystem_nvme_create: %v", "EOF"),
-			true,
-			false,
-		},
-		"valid request with ID mismatch SPDK response": {
-			testSubsystemID,
-			&pb.NvmeSubsystem{
-				Spec: spec,
-			},
-			nil,
-			[]string{`{"id":0,"error":{"code":0,"message":""},"result":false}`},
-			codes.Unknown,
-			fmt.Sprintf("subsystem_nvme_create: %v", "json response ID mismatch"),
-			true,
-			false,
-		},
-		"valid request with error code from SPDK response": {
-			testSubsystemID,
-			&pb.NvmeSubsystem{
-				Spec: spec,
-			},
-			nil,
-			[]string{`{"id":%d,"error":{"code":1,"message":"myopierr"},"result":false}`},
-			codes.Unknown,
-			fmt.Sprintf("subsystem_nvme_create: %v", "json response error: myopierr"),
-			true,
-			false,
-		},
-		"valid request with error code from SPDK version response": {
-			testSubsystemID,
-			&pb.NvmeSubsystem{
-				Spec: spec,
-			},
-			nil,
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":true}`, `{"id":%d,"error":{"code":1,"message":"myopierr"},"result":false}`},
-			codes.Unknown,
-			fmt.Sprintf("spdk_get_version: %v", "json response error: myopierr"),
-			true,
-			false,
-		},
-		"valid request with valid SPDK response": {
-			testSubsystemID,
-			&pb.NvmeSubsystem{
-				Spec: spec,
-			},
-			&pb.NvmeSubsystem{
-				Name: testSubsystemName,
-				Spec: &pb.NvmeSubsystemSpec{
-					Nqn:          "nqn.2022-09.io.spdk:opi3",
-					SerialNumber: "OpiSerialNumber",
-					ModelNumber:  "OpiModelNumber",
-				},
-				Status: &pb.NvmeSubsystemStatus{
-					FirmwareRevision: "SPDK v20.10",
-				},
-			},
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":true}`,
-				`{"jsonrpc":"2.0","id":%d,"result":{"version":"SPDK v20.10","fields":{"major":20,"minor":10,"patch":0,"suffix":""}}}`},
-			codes.OK,
-			"",
-			true,
-			false,
-		},
-		"already exists": {
-			testSubsystemID,
-			&pb.NvmeSubsystem{
-				Spec: spec,
-			},
-			&testSubsystem,
-			[]string{""},
-			codes.OK,
-			"",
-			false,
-			true,
-		},
-	}
-
-	// run tests
-	for name, tt := range tests {
-		t.Run(name, func(t *testing.T) {
-			testEnv := createTestEnvironment(tt.start, tt.spdk)
-			defer testEnv.Close()
-
-			testEnv.opiSpdkServer.Controllers[testControllerName] = &testController
-			testEnv.opiSpdkServer.Namespaces[testNamespaceName] = &testNamespace
-			if tt.exist {
-				testEnv.opiSpdkServer.Subsystems[testSubsystemName] = &testSubsystem
-				testEnv.opiSpdkServer.Subsystems[testSubsystemName].Name = testSubsystemName
-			}
-			if tt.out != nil {
-				tt.out.Name = testSubsystemName
-			}
-
-			request := &pb.CreateNvmeSubsystemRequest{NvmeSubsystem: tt.in, NvmeSubsystemId: tt.id}
-			response, err := testEnv.client.CreateNvmeSubsystem(testEnv.ctx, request)
-			if response != nil {
-				mtt, _ := proto.Marshal(tt.out)
-				mResponse, _ := proto.Marshal(response)
-				if !bytes.Equal(mtt, mResponse) {
-					t.Error("response: expected", tt.out, "received", response)
-				}
-			}
-
-			if err != nil {
-				if er, ok := status.FromError(err); ok {
-					if er.Code() != tt.errCode {
-						t.Error("error code: expected", tt.errCode, "received", er.Code())
-					}
-					if er.Message() != tt.errMsg {
-						t.Error("error message: expected", tt.errMsg, "received", er.Message())
-					}
-				}
-			}
-		})
-	}
-}
-
-func TestFrontEnd_UpdateNvmeSubsystem(t *testing.T) {
-	tests := map[string]struct {
-		mask    *fieldmaskpb.FieldMask
-		in      *pb.NvmeSubsystem
-		out     *pb.NvmeSubsystem
-		spdk    []string
-		errCode codes.Code
-		errMsg  string
-		start   bool
-	}{
-		"invalid fieldmask": {
-			&fieldmaskpb.FieldMask{Paths: []string{"*", "author"}},
-			&pb.NvmeSubsystem{
-				Name: testSubsystemName,
-			},
-			nil,
-			[]string{""},
-			codes.Unknown,
-			fmt.Sprintf("invalid field path: %s", "'*' must not be used with other paths"),
-			false,
-		},
-		"unimplemented method": {
-			nil,
-			&pb.NvmeSubsystem{
-				Name: testSubsystemName,
-			},
-			nil,
-			[]string{""},
-			codes.Unimplemented,
-			fmt.Sprintf("%v method is not implemented", "UpdateNvmeSubsystem"),
-			false,
-		},
-		"valid request with unknown key": {
-			nil,
-			&pb.NvmeSubsystem{
-				Name: server.ResourceIDToVolumeName("unknown-id"),
-				Spec: &pb.NvmeSubsystemSpec{
-					Nqn: "nqn.2022-09.io.spdk:opi3",
-				},
-			},
-			nil,
-			[]string{""},
-			codes.NotFound,
-			fmt.Sprintf("unable to find key %v", server.ResourceIDToVolumeName("unknown-id")),
-			false,
-		},
-	}
-
-	// run tests
-	for name, tt := range tests {
-		t.Run(name, func(t *testing.T) {
-			testEnv := createTestEnvironment(tt.start, tt.spdk)
-			defer testEnv.Close()
-
-			testEnv.opiSpdkServer.Subsystems[testSubsystemName] = &testSubsystem
-			testEnv.opiSpdkServer.Controllers[testControllerName] = &testController
-			testEnv.opiSpdkServer.Namespaces[testNamespaceName] = &testNamespace
-
-			request := &pb.UpdateNvmeSubsystemRequest{NvmeSubsystem: tt.in, UpdateMask: tt.mask}
-			response, err := testEnv.client.UpdateNvmeSubsystem(testEnv.ctx, request)
-			if response != nil {
-				t.Error("response: expected", tt.out, "received", response)
-			}
-
-			if err != nil {
-				if er, ok := status.FromError(err); ok {
-					if er.Code() != tt.errCode {
-						t.Error("error code: expected", tt.errCode, "received", er.Code())
-					}
-					if er.Message() != tt.errMsg {
-						t.Error("error message: expected", tt.errMsg, "received", er.Message())
-					}
-				}
-			}
-		})
-	}
-}
-
-func TestFrontEnd_ListNvmeSubsystem(t *testing.T) {
-	tests := map[string]struct {
-		out     []*pb.NvmeSubsystem
-		spdk    []string
-		errCode codes.Code
-		errMsg  string
-		start   bool
-		size    int32
-		token   string
-	}{
-		"valid request with invalid SPDK response": {
-			nil,
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":[]}`},
-			codes.InvalidArgument,
-			fmt.Sprintf("Could not create NQN: %v", "nqn.2022-09.io.spdk:opi3"),
-			true,
-			0,
-			"",
-		},
-		"valid request with empty SPDK response": {
-			nil,
-			[]string{""},
-			codes.Unknown,
-			fmt.Sprintf("subsystem_nvme_list: %v", "EOF"),
-			true,
-			0,
-			"",
-		},
-		"valid request with ID mismatch SPDK response": {
-			nil,
-			[]string{`{"id":0,"error":{"code":0,"message":""},"result":[]}`},
-			codes.Unknown,
-			fmt.Sprintf("subsystem_nvme_list: %v", "json response ID mismatch"),
-			true,
-			0,
-			"",
-		},
-		"valid request with error code from SPDK response": {
-			nil,
-			[]string{`{"id":%d,"error":{"code":1,"message":"myopierr"},"result":[]}`},
-			codes.Unknown,
-			fmt.Sprintf("subsystem_nvme_list: %v", "json response error: myopierr"),
-			true,
-			0,
-			"",
-		},
-		"pagination negative": {
-			nil,
-			[]string{},
-			codes.InvalidArgument,
-			"negative PageSize is not allowed",
-			false,
-			-10,
-			"",
-		},
-		"pagination error": {
-			nil,
-			[]string{},
-			codes.NotFound,
-			fmt.Sprintf("unable to find pagination token %s", "unknown-pagination-token"),
-			false,
-			0,
-			"unknown-pagination-token",
-		},
-		"pagination": {
-			[]*pb.NvmeSubsystem{
-				{
-					Spec: &pb.NvmeSubsystemSpec{
-						Nqn:          "nqn.2022-09.io.spdk:opi1",
-						SerialNumber: "OpiSerialNumber1",
-						ModelNumber:  "OpiModelNumber1",
-					},
-				},
-			},
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":[{"nqn": "nqn.2022-09.io.spdk:opi1", "serial_number": "OpiSerialNumber1", "model_number": "OpiModelNumber1"},{"nqn": "nqn.2022-09.io.spdk:opi2", "serial_number": "OpiSerialNumber2", "model_number": "OpiModelNumber2"},{"nqn": "nqn.2022-09.io.spdk:opi3", "serial_number": "OpiSerialNumber3", "model_number": "OpiModelNumber3"}]}`},
-			codes.OK,
-			"",
-			true,
-			1,
-			"",
-		},
-		"pagination overflow": {
-			[]*pb.NvmeSubsystem{
-				{
-					Spec: &pb.NvmeSubsystemSpec{
-						Nqn:          "nqn.2022-09.io.spdk:opi1",
-						SerialNumber: "OpiSerialNumber1",
-						ModelNumber:  "OpiModelNumber1",
-					},
-				},
-				{
-					Spec: &pb.NvmeSubsystemSpec{
-						Nqn:          "nqn.2022-09.io.spdk:opi2",
-						SerialNumber: "OpiSerialNumber2",
-						ModelNumber:  "OpiModelNumber2",
-					},
-				},
-				{
-					Spec: &pb.NvmeSubsystemSpec{
-						Nqn:          "nqn.2022-09.io.spdk:opi3",
-						SerialNumber: "OpiSerialNumber3",
-						ModelNumber:  "OpiModelNumber3",
-					},
-				},
-			},
-			// {'jsonrpc': '2.0', 'id': 1, 'result': [{'nqn': 'nqn.2020-12.mlnx.snap', 'serial_number': 'Mellanox_Nvme_SNAP', 'model_number': 'Mellanox Nvme SNAP Controller', 'controllers': [{'name': 'NvmeEmu0pf1', 'cntlid': 0, 'pci_bdf': 'ca:00.3', 'pci_index': 1}]}]}
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":[{"nqn": "nqn.2022-09.io.spdk:opi1", "serial_number": "OpiSerialNumber1", "model_number": "OpiModelNumber1"},{"nqn": "nqn.2022-09.io.spdk:opi2", "serial_number": "OpiSerialNumber2", "model_number": "OpiModelNumber2"},{"nqn": "nqn.2022-09.io.spdk:opi3", "serial_number": "OpiSerialNumber3", "model_number": "OpiModelNumber3"}]}`},
-			codes.OK,
-			"",
-			true,
-			1000,
-			"",
-		},
-		"pagination offset": {
-			[]*pb.NvmeSubsystem{
-				{
-					Spec: &pb.NvmeSubsystemSpec{
-						Nqn:          "nqn.2022-09.io.spdk:opi2",
-						SerialNumber: "OpiSerialNumber2",
-						ModelNumber:  "OpiModelNumber2",
-					},
-				},
-			},
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":[{"nqn": "nqn.2022-09.io.spdk:opi1", "serial_number": "OpiSerialNumber1", "model_number": "OpiModelNumber1"},{"nqn": "nqn.2022-09.io.spdk:opi2", "serial_number": "OpiSerialNumber2", "model_number": "OpiModelNumber2"},{"nqn": "nqn.2022-09.io.spdk:opi3", "serial_number": "OpiSerialNumber3", "model_number": "OpiModelNumber3"}]}`},
-			codes.OK,
-			"",
-			true,
-			1,
-			"existing-pagination-token",
-		},
-		"valid request with valid SPDK response": {
-			[]*pb.NvmeSubsystem{
-				{
-					Spec: &pb.NvmeSubsystemSpec{
-						Nqn:          "nqn.2022-09.io.spdk:opi1",
-						SerialNumber: "OpiSerialNumber1",
-						ModelNumber:  "OpiModelNumber1",
-					},
-				},
-				{
-					Spec: &pb.NvmeSubsystemSpec{
-						Nqn:          "nqn.2022-09.io.spdk:opi2",
-						SerialNumber: "OpiSerialNumber2",
-						ModelNumber:  "OpiModelNumber2",
-					},
-				},
-				{
-					Spec: &pb.NvmeSubsystemSpec{
-						Nqn:          "nqn.2022-09.io.spdk:opi3",
-						SerialNumber: "OpiSerialNumber3",
-						ModelNumber:  "OpiModelNumber3",
-					},
-				},
-			},
-			// {'jsonrpc': '2.0', 'id': 1, 'result': [{'nqn': 'nqn.2020-12.mlnx.snap', 'serial_number': 'Mellanox_Nvme_SNAP', 'model_number': 'Mellanox Nvme SNAP Controller', 'controllers': [{'name': 'NvmeEmu0pf1', 'cntlid': 0, 'pci_bdf': 'ca:00.3', 'pci_index': 1}]}]}
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":[{"nqn": "nqn.2022-09.io.spdk:opi1", "serial_number": "OpiSerialNumber1", "model_number": "OpiModelNumber1"},{"nqn": "nqn.2022-09.io.spdk:opi2", "serial_number": "OpiSerialNumber2", "model_number": "OpiModelNumber2"},{"nqn": "nqn.2022-09.io.spdk:opi3", "serial_number": "OpiSerialNumber3", "model_number": "OpiModelNumber3"}]}`},
-			codes.OK,
-			"",
-			true,
-			0,
-			"",
-		},
-	}
-
-	// run tests
-	for name, tt := range tests {
-		t.Run(name, func(t *testing.T) {
-			testEnv := createTestEnvironment(tt.start, tt.spdk)
-			defer testEnv.Close()
-
-			testEnv.opiSpdkServer.Subsystems[testSubsystemName] = &testSubsystem
-			testEnv.opiSpdkServer.Controllers[testControllerName] = &testController
-			testEnv.opiSpdkServer.Namespaces[testNamespaceName] = &testNamespace
-			testEnv.opiSpdkServer.Pagination["existing-pagination-token"] = 1
-
-			request := &pb.ListNvmeSubsystemsRequest{PageSize: tt.size, PageToken: tt.token}
-			response, err := testEnv.client.ListNvmeSubsystems(testEnv.ctx, request)
-			if response != nil {
-				if !reflect.DeepEqual(response.NvmeSubsystems, tt.out) {
-					t.Error("response: expected", tt.out, "received", response.NvmeSubsystems)
-				}
-			}
-
-			if err != nil {
-				if er, ok := status.FromError(err); ok {
-					if er.Code() != tt.errCode {
-						t.Error("error code: expected", tt.errCode, "received", er.Code())
-					}
-					if er.Message() != tt.errMsg {
-						t.Error("error message: expected", tt.errMsg, "received", er.Message())
-					}
-				}
-			}
-		})
-	}
-}
-
-func TestFrontEnd_GetNvmeSubsystem(t *testing.T) {
-	tests := map[string]struct {
-		in      string
-		out     *pb.NvmeSubsystem
-		spdk    []string
-		errCode codes.Code
-		errMsg  string
-		start   bool
-	}{
-		"valid request with invalid SPDK response": {
-			testSubsystemName,
-			nil,
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":[]}`},
-			codes.InvalidArgument,
-			fmt.Sprintf("Could not find NQN: %v", "nqn.2022-09.io.spdk:opi3"),
-			true,
-		},
-		"valid request with empty SPDK response": {
-			testSubsystemName,
-			nil,
-			[]string{""},
-			codes.Unknown,
-			fmt.Sprintf("subsystem_nvme_list: %v", "EOF"),
-			true,
-		},
-		"valid request with ID mismatch SPDK response": {
-			testSubsystemName,
-			nil,
-			[]string{`{"id":0,"error":{"code":0,"message":""},"result":[]}`},
-			codes.Unknown,
-			fmt.Sprintf("subsystem_nvme_list: %v", "json response ID mismatch"),
-			true,
-		},
-		"valid request with error code from SPDK response": {
-			testSubsystemName,
-			nil,
-			[]string{`{"id":%d,"error":{"code":1,"message":"myopierr"},"result":[]}`},
-			codes.Unknown,
-			fmt.Sprintf("subsystem_nvme_list: %v", "json response error: myopierr"),
-			true,
-		},
-		"valid request with valid SPDK response": {
-			testSubsystemName,
-			&pb.NvmeSubsystem{
-				Spec: &pb.NvmeSubsystemSpec{
-					Nqn:          "nqn.2022-09.io.spdk:opi3",
-					SerialNumber: "OpiSerialNumber3",
-					ModelNumber:  "OpiModelNumber3",
-				},
-				Status: &pb.NvmeSubsystemStatus{
-					FirmwareRevision: "TBD",
-				},
-			},
-			// {'jsonrpc': '2.0', 'id': 1, 'result': [{'nqn': 'nqn.2020-12.mlnx.snap', 'serial_number': 'Mellanox_Nvme_SNAP', 'model_number': 'Mellanox Nvme SNAP Controller', 'controllers': [{'name': 'NvmeEmu0pf1', 'cntlid': 0, 'pci_bdf': 'ca:00.3', 'pci_index': 1}]}]}
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":[{"nqn": "nqn.2022-09.io.spdk:opi1", "serial_number": "OpiSerialNumber1", "model_number": "OpiModelNumber1"},{"nqn": "nqn.2022-09.io.spdk:opi2", "serial_number": "OpiSerialNumber2", "model_number": "OpiModelNumber2"},{"nqn": "nqn.2022-09.io.spdk:opi3", "serial_number": "OpiSerialNumber3", "model_number": "OpiModelNumber3"}]}`},
-			codes.OK,
-			"",
-			true,
-		},
-		"valid request with unknown key": {
-			server.ResourceIDToVolumeName("unknown-subsystem-id"),
-			nil,
-			[]string{""},
-			codes.NotFound,
-			fmt.Sprintf("unable to find key %v", server.ResourceIDToVolumeName("unknown-subsystem-id")),
-			false,
-		},
-	}
-
-	// run tests
-	for name, tt := range tests {
-		t.Run(name, func(t *testing.T) {
-			testEnv := createTestEnvironment(tt.start, tt.spdk)
-			defer testEnv.Close()
-
-			testEnv.opiSpdkServer.Subsystems[testSubsystemName] = &testSubsystem
-			testEnv.opiSpdkServer.Controllers[testControllerName] = &testController
-			testEnv.opiSpdkServer.Namespaces[testNamespaceName] = &testNamespace
-
-			request := &pb.GetNvmeSubsystemRequest{Name: tt.in}
-			response, err := testEnv.client.GetNvmeSubsystem(testEnv.ctx, request)
-			if response != nil {
-				if !reflect.DeepEqual(response.Spec, tt.out.Spec) {
-					t.Error("response: expected", tt.out.GetSpec(), "received", response.GetSpec())
-				}
-				if !reflect.DeepEqual(response.Status, tt.out.Status) {
-					t.Error("response: expected", tt.out.GetStatus(), "received", response.GetStatus())
-				}
-			}
-
-			if err != nil {
-				if er, ok := status.FromError(err); ok {
-					if er.Code() != tt.errCode {
-						t.Error("error code: expected", tt.errCode, "received", er.Code())
-					}
-					if er.Message() != tt.errMsg {
-						t.Error("error message: expected", tt.errMsg, "received", er.Message())
-					}
-				}
-			}
-		})
-	}
-}
-
-func TestFrontEnd_NvmeSubsystemStats(t *testing.T) {
-	tests := map[string]struct {
-		in      string
-		out     *pb.NvmeSubsystemStatsResponse
-		spdk    []string
-		errCode codes.Code
-		errMsg  string
-		start   bool
-	}{
-		"unimplemented method": {
-			testSubsystemName,
-			nil,
-			[]string{""},
-			codes.Unimplemented,
-			fmt.Sprintf("%v method is not implemented", "UpdateNvmeSubsystem"),
-			false,
-		},
-	}
-
-	// run tests
-	for name, tt := range tests {
-		t.Run(name, func(t *testing.T) {
-			testEnv := createTestEnvironment(tt.start, tt.spdk)
-			defer testEnv.Close()
-
-			testEnv.opiSpdkServer.Subsystems[testSubsystemName] = &testSubsystem
-			testEnv.opiSpdkServer.Controllers[testControllerName] = &testController
-			testEnv.opiSpdkServer.Namespaces[testNamespaceName] = &testNamespace
-
-			request := &pb.NvmeSubsystemStatsRequest{SubsystemId: &pc.ObjectKey{Value: tt.in}}
-			response, err := testEnv.client.NvmeSubsystemStats(testEnv.ctx, request)
-			if response != nil {
-				t.Error("response: expected", tt.out, "received", response)
-			}
-
-			if err != nil {
-				if er, ok := status.FromError(err); ok {
-					if er.Code() != tt.errCode {
-						t.Error("error code: expected", tt.errCode, "received", er.Code())
-					}
-					if er.Message() != tt.errMsg {
-						t.Error("error message: expected", tt.errMsg, "received", er.Message())
-					}
-				}
-			}
-		})
-	}
-}
-
-func TestFrontEnd_CreateNvmeController(t *testing.T) {
-	spec := &pb.NvmeControllerSpec{
-		SubsystemId:      &pc.ObjectKey{Value: testSubsystemName},
-		PcieId:           &pb.PciEndpoint{PhysicalFunction: 1, VirtualFunction: 2},
-		NvmeControllerId: 1,
-	}
-	controllerSpec := &pb.NvmeControllerSpec{
-		SubsystemId:      &pc.ObjectKey{Value: testSubsystemName},
-		PcieId:           &pb.PciEndpoint{PhysicalFunction: 1, VirtualFunction: 2},
-		NvmeControllerId: 17,
-	}
-	tests := map[string]struct {
-		id      string
-		in      *pb.NvmeController
-		out     *pb.NvmeController
-		spdk    []string
-		errCode codes.Code
-		errMsg  string
-		start   bool
-		exist   bool
-	}{
-		"illegal resource_id": {
-			"CapitalLettersNotAllowed",
-			&pb.NvmeController{
-				Spec: &pb.NvmeControllerSpec{
-					SubsystemId:      &pc.ObjectKey{Value: testSubsystemName},
-					PcieId:           &pb.PciEndpoint{PhysicalFunction: 1, VirtualFunction: 2},
-					NvmeControllerId: 1,
-				},
-			},
-			nil,
-			[]string{""},
-			codes.Unknown,
-			fmt.Sprintf("user-settable ID must only contain lowercase, numbers and hyphens (%v)", "got: 'C' in position 0"),
-			false,
-			false,
-		},
-		"valid request with invalid SPDK response": {
-			testControllerID,
-			&pb.NvmeController{
-				Spec: spec,
-			},
-			nil,
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":{"name": "NvmeEmu0pf0", "cntlid": -1}}`},
-			codes.InvalidArgument,
-			fmt.Sprintf("Could not create CTRL: %v", testControllerName),
-			true,
-			false,
-		},
-		"valid request with empty SPDK response": {
-			testControllerID,
-			&pb.NvmeController{
-				Spec: spec,
-			},
-			nil,
-			[]string{""},
-			codes.Unknown,
-			fmt.Sprintf("controller_nvme_create: %v", "EOF"),
-			true,
-			false,
-		},
-		"valid request with ID mismatch SPDK response": {
-			testControllerID,
-			&pb.NvmeController{
-				Spec: spec,
-			},
-			nil,
-			[]string{`{"id":0,"error":{"code":0,"message":""},"result":{"name": "NvmeEmu0pf0", "cntlid": 17}}`},
-			codes.Unknown,
-			fmt.Sprintf("controller_nvme_create: %v", "json response ID mismatch"),
-			true,
-			false,
-		},
-		"valid request with error code from SPDK response": {
-			testControllerID,
-			&pb.NvmeController{
-				Spec: spec,
-			},
-			nil,
-			[]string{`{"id":%d,"error":{"code":-32602,"message":"Invalid parameters"}}`},
-			codes.Unknown,
-			fmt.Sprintf("controller_nvme_create: %v", "json response error: Invalid parameters"),
-			true,
-			false,
-		},
-		"valid request with valid SPDK response": {
-			testControllerID,
-			&pb.NvmeController{
-				Spec: controllerSpec,
-			},
-			&pb.NvmeController{
-				Name: testControllerName,
-				Spec: &pb.NvmeControllerSpec{
-					SubsystemId:      &pc.ObjectKey{Value: testSubsystemName},
-					PcieId:           &pb.PciEndpoint{PhysicalFunction: 1, VirtualFunction: 2},
-					NvmeControllerId: 17,
-				},
-				Status: &pb.NvmeControllerStatus{
-					Active: true,
-				},
-			},
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":{"name": "NvmeEmu0pf0", "cntlid": 17}}`},
-			codes.OK,
-			"",
-			true,
-			false,
-		},
-		"already exists": {
-			testControllerID,
-			&pb.NvmeController{
-				Spec: spec,
-			},
-			&testController,
-			[]string{""},
-			codes.OK,
-			"",
-			false,
-			true,
-		},
-	}
-
-	// run tests
-	for name, tt := range tests {
-		t.Run(name, func(t *testing.T) {
-			testEnv := createTestEnvironment(tt.start, tt.spdk)
-			defer testEnv.Close()
-
-			testEnv.opiSpdkServer.Subsystems[testSubsystemName] = &testSubsystem
-			testEnv.opiSpdkServer.Namespaces[testNamespaceName] = &testNamespace
-			if tt.exist {
-				testEnv.opiSpdkServer.Controllers[testControllerName] = &testController
-				testEnv.opiSpdkServer.Controllers[testControllerName].Name = testControllerName
-			}
-			if tt.out != nil {
-				tt.out.Name = testControllerName
-			}
-
-			request := &pb.CreateNvmeControllerRequest{NvmeController: tt.in, NvmeControllerId: tt.id}
-			response, err := testEnv.client.CreateNvmeController(testEnv.ctx, request)
-			if response != nil {
-				mtt, _ := proto.Marshal(tt.out)
-				mResponse, _ := proto.Marshal(response)
-				if !bytes.Equal(mtt, mResponse) {
-					t.Error("response: expected", tt.out, "received", response)
-				}
-			}
-
-			if err != nil {
-				if er, ok := status.FromError(err); ok {
-					if er.Code() != tt.errCode {
-						t.Error("error code: expected", tt.errCode, "received", er.Code())
-					}
-					if er.Message() != tt.errMsg {
-						t.Error("error message: expected", tt.errMsg, "received", er.Message())
-					}
-				}
-			}
-		})
-	}
-}
-
-func TestFrontEnd_UpdateNvmeController(t *testing.T) {
-	tests := map[string]struct {
-		mask    *fieldmaskpb.FieldMask
-		in      *pb.NvmeController
-		out     *pb.NvmeController
-		spdk    []string
-		errCode codes.Code
-		errMsg  string
-		start   bool
-	}{
-		"invalid fieldmask": {
-			&fieldmaskpb.FieldMask{Paths: []string{"*", "author"}},
-			&pb.NvmeController{
-				Name: testControllerName,
-			},
-			nil,
-			[]string{""},
-			codes.Unknown,
-			fmt.Sprintf("invalid field path: %s", "'*' must not be used with other paths"),
-			false,
-		},
-		"unimplemented method": {
-			nil,
-			&pb.NvmeController{
-				Name: testControllerName,
-			},
-			nil,
-			[]string{""},
-			codes.Unimplemented,
-			fmt.Sprintf("%v method is not implemented", "UpdateNvmeController"),
-			false,
-		},
-		"valid request with unknown key": {
-			nil,
-			&pb.NvmeController{
-				Name: server.ResourceIDToVolumeName("unknown-id"),
-			},
-			nil,
-			[]string{""},
-			codes.NotFound,
-			fmt.Sprintf("unable to find key %v", server.ResourceIDToVolumeName("unknown-id")),
-			false,
-		},
-	}
-
-	// run tests
-	for name, tt := range tests {
-		t.Run(name, func(t *testing.T) {
-			testEnv := createTestEnvironment(tt.start, tt.spdk)
-			defer testEnv.Close()
-
-			testEnv.opiSpdkServer.Subsystems[testSubsystemName] = &testSubsystem
-			testEnv.opiSpdkServer.Controllers[testControllerName] = &testController
-			testEnv.opiSpdkServer.Namespaces[testNamespaceName] = &testNamespace
-
-			request := &pb.UpdateNvmeControllerRequest{NvmeController: tt.in, UpdateMask: tt.mask}
-			response, err := testEnv.client.UpdateNvmeController(testEnv.ctx, request)
-			if response != nil {
-				t.Error("response: expected", tt.out, "received", response)
-			}
-
-			if err != nil {
-				if er, ok := status.FromError(err); ok {
-					if er.Code() != tt.errCode {
-						t.Error("error code: expected", tt.errCode, "received", er.Code())
-					}
-					if er.Message() != tt.errMsg {
-						t.Error("error message: expected", tt.errMsg, "received", er.Message())
-					}
-				}
-			}
-		})
-	}
-}
-
-func TestFrontEnd_ListNvmeControllers(t *testing.T) {
-	tests := map[string]struct {
-		in      string
-		out     []*pb.NvmeController
-		spdk    []string
-		errCode codes.Code
-		errMsg  string
-		start   bool
-		size    int32
-		token   string
-	}{
-		"valid request with invalid SPDK response": {
-			testSubsystemName,
-			nil,
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":[]}`},
-			codes.InvalidArgument,
-			fmt.Sprintf("Could not create NQN: %v", "nqn.2022-09.io.spdk:opi3"),
-			true,
-			0,
-			"",
-		},
-		"valid request with empty SPDK response": {
-			testSubsystemName,
-			nil,
-			[]string{""},
-			codes.Unknown,
-			fmt.Sprintf("controller_list: %v", "EOF"),
-			true,
-			0,
-			"",
-		},
-		"valid request with ID mismatch SPDK response": {
-			testSubsystemName,
-			nil,
-			[]string{`{"id":0,"error":{"code":0,"message":""},"result":[]}`},
-			codes.Unknown,
-			fmt.Sprintf("controller_list: %v", "json response ID mismatch"),
-			true,
-			0,
-			"",
-		},
-		"valid request with error code from SPDK response": {
-			testSubsystemName,
-			nil,
-			[]string{`{"id":%d,"error":{"code":1,"message":"myopierr"},"result":[]}`},
-			codes.Unknown,
-			fmt.Sprintf("controller_list: %v", "json response error: myopierr"),
-			true,
-			0,
-			"",
-		},
-		"pagination negative": {
-			testSubsystemName,
-			nil,
-			[]string{},
-			codes.InvalidArgument,
-			"negative PageSize is not allowed",
-			false,
-			-10,
-			"",
-		},
-		"pagination error": {
-			testSubsystemName,
-			nil,
-			[]string{},
-			codes.NotFound,
-			fmt.Sprintf("unable to find pagination token %s", "unknown-pagination-token"),
-			false,
-			0,
-			"unknown-pagination-token",
-		},
-		"pagination": {
-			testSubsystemName,
-			[]*pb.NvmeController{
-				{
-					Spec: &pb.NvmeControllerSpec{
-						NvmeControllerId: 1,
-					},
-				},
-			},
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":[{"subnqn": "nqn.2022-09.io.spdk:opi3", "cntlid": 1, "name": "NvmeEmu0pf1", "type": "nvme", "pci_index": 1, "pci_bdf": "ca:00.3"},{"subnqn": "nqn.2022-09.io.spdk:opi3", "cntlid": 2, "name": "NvmeEmu0pf1", "type": "nvme", "pci_index": 2, "pci_bdf": "ca:00.4"},{"subnqn": "nqn.2022-09.io.spdk:opi3", "cntlid": 3, "name": "NvmeEmu0pf1", "type": "nvme", "pci_index": 3, "pci_bdf": "ca:00.5"}]}`},
-			codes.OK,
-			"",
-			true,
-			1,
-			"",
-		},
-		"pagination overflow": {
-			testSubsystemName,
-			[]*pb.NvmeController{
-				{
-					Spec: &pb.NvmeControllerSpec{
-						NvmeControllerId: 1,
-					},
-				},
-				{
-					Spec: &pb.NvmeControllerSpec{
-						NvmeControllerId: 2,
-					},
-				},
-				{
-					Spec: &pb.NvmeControllerSpec{
-						NvmeControllerId: 3,
-					},
-				},
-			},
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":[{"subnqn": "nqn.2022-09.io.spdk:opi3", "cntlid": 1, "name": "NvmeEmu0pf1", "type": "nvme", "pci_index": 1, "pci_bdf": "ca:00.3"},{"subnqn": "nqn.2022-09.io.spdk:opi3", "cntlid": 2, "name": "NvmeEmu0pf1", "type": "nvme", "pci_index": 2, "pci_bdf": "ca:00.4"},{"subnqn": "nqn.2022-09.io.spdk:opi3", "cntlid": 3, "name": "NvmeEmu0pf1", "type": "nvme", "pci_index": 3, "pci_bdf": "ca:00.5"}]}`},
-			codes.OK,
-			"",
-			true,
-			1000,
-			"",
-		},
-		"pagination offset": {
-			testSubsystemName,
-			[]*pb.NvmeController{
-				{
-					Spec: &pb.NvmeControllerSpec{
-						NvmeControllerId: 2,
-					},
-				},
-			},
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":[{"subnqn": "nqn.2022-09.io.spdk:opi3", "cntlid": 1, "name": "NvmeEmu0pf1", "type": "nvme", "pci_index": 1, "pci_bdf": "ca:00.3"},{"subnqn": "nqn.2022-09.io.spdk:opi3", "cntlid": 2, "name": "NvmeEmu0pf1", "type": "nvme", "pci_index": 2, "pci_bdf": "ca:00.4"},{"subnqn": "nqn.2022-09.io.spdk:opi3", "cntlid": 3, "name": "NvmeEmu0pf1", "type": "nvme", "pci_index": 3, "pci_bdf": "ca:00.5"}]}`},
-			codes.OK,
-			"",
-			true,
-			1,
-			"existing-pagination-token",
-		},
-		"valid request with valid SPDK response": {
-			testSubsystemName,
-			[]*pb.NvmeController{
-				{
-					Spec: &pb.NvmeControllerSpec{
-						NvmeControllerId: 1,
-					},
-				},
-				{
-					Spec: &pb.NvmeControllerSpec{
-						NvmeControllerId: 2,
-					},
-				},
-				{
-					Spec: &pb.NvmeControllerSpec{
-						NvmeControllerId: 3,
-					},
-				},
-			},
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":[{"subnqn": "nqn.2022-09.io.spdk:opi3", "cntlid": 1, "name": "NvmeEmu0pf1", "type": "nvme", "pci_index": 1, "pci_bdf": "ca:00.3"},{"subnqn": "nqn.2022-09.io.spdk:opi3", "cntlid": 2, "name": "NvmeEmu0pf1", "type": "nvme", "pci_index": 2, "pci_bdf": "ca:00.4"},{"subnqn": "nqn.2022-09.io.spdk:opi3", "cntlid": 3, "name": "NvmeEmu0pf1", "type": "nvme", "pci_index": 3, "pci_bdf": "ca:00.5"}]}`},
-			codes.OK,
-			"",
-			true,
-			0,
-			"",
-		},
-		"valid request with unknown key": {
-			server.ResourceIDToVolumeName("unknown-subsystem-id"),
-			nil,
-			[]string{""},
-			codes.NotFound,
-			fmt.Sprintf("unable to find key %v", server.ResourceIDToVolumeName("unknown-subsystem-id")),
-			false,
-			0,
-			"",
-		},
-	}
-
-	// run tests
-	for name, tt := range tests {
-		t.Run(name, func(t *testing.T) {
-			testEnv := createTestEnvironment(tt.start, tt.spdk)
-			defer testEnv.Close()
-
-			testEnv.opiSpdkServer.Subsystems[testSubsystemName] = &testSubsystem
-			testEnv.opiSpdkServer.Controllers[testControllerName] = &testController
-			testEnv.opiSpdkServer.Namespaces[testNamespaceName] = &testNamespace
-			testEnv.opiSpdkServer.Pagination["existing-pagination-token"] = 1
-
-			request := &pb.ListNvmeControllersRequest{Parent: tt.in, PageSize: tt.size, PageToken: tt.token}
-			response, err := testEnv.client.ListNvmeControllers(testEnv.ctx, request)
-			if response != nil {
-				if !reflect.DeepEqual(response.NvmeControllers, tt.out) {
-					t.Error("response: expected", tt.out, "received", response.NvmeControllers)
-				}
-			}
-
-			if err != nil {
-				if er, ok := status.FromError(err); ok {
-					if er.Code() != tt.errCode {
-						t.Error("error code: expected", tt.errCode, "received", er.Code())
-					}
-					if er.Message() != tt.errMsg {
-						t.Error("error message: expected", tt.errMsg, "received", er.Message())
-					}
-				}
-			}
-		})
-	}
-}
-
-func TestFrontEnd_GetNvmeController(t *testing.T) {
-	tests := map[string]struct {
-		in      string
-		out     *pb.NvmeController
-		spdk    []string
-		errCode codes.Code
-		errMsg  string
-		start   bool
-	}{
-		"valid request with invalid SPDK response": {
-			testControllerName,
-			nil,
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":[]}`},
-			codes.InvalidArgument,
-			fmt.Sprintf("Could not find NvmeControllerId: %v", "17"),
-			true,
-		},
-		"valid request with empty SPDK response": {
-			testControllerName,
-			nil,
-			[]string{""},
-			codes.Unknown,
-			fmt.Sprintf("controller_list: %v", "EOF"),
-			true,
-		},
-		"valid request with ID mismatch SPDK response": {
-			testControllerName,
-			nil,
-			[]string{`{"id":0,"error":{"code":0,"message":""},"result":[]}`},
-			codes.Unknown,
-			fmt.Sprintf("controller_list: %v", "json response ID mismatch"),
-			true,
-		},
-		"valid request with error code from SPDK response": {
-			testControllerName,
-			nil,
-			[]string{`{"id":%d,"error":{"code":1,"message":"myopierr"},"result":[]}`},
-			codes.Unknown,
-			fmt.Sprintf("controller_list: %v", "json response error: myopierr"),
-			true,
-		},
-		"valid request with valid SPDK response": {
-			testControllerName,
-			&pb.NvmeController{
-				Spec: &pb.NvmeControllerSpec{
-					NvmeControllerId: 17,
-				},
-				Status: &pb.NvmeControllerStatus{
-					Active: true,
-				},
-			},
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":[{"subnqn": "nqn.2022-09.io.spdk:opi3", "cntlid": 1, "name": "NvmeEmu0pf1", "type": "nvme", "pci_index": 1, "pci_bdf": "ca:00.3"},{"subnqn": "nqn.2022-09.io.spdk:opi3", "cntlid": 17, "name": "NvmeEmu0pf1", "type": "nvme", "pci_index": 2, "pci_bdf": "ca:00.4"},{"subnqn": "nqn.2022-09.io.spdk:opi3", "cntlid": 3, "name": "NvmeEmu0pf1", "type": "nvme", "pci_index": 3, "pci_bdf": "ca:00.5"}]}`},
-			codes.OK,
-			"",
-			true,
-		},
-		"valid request with unknown key": {
-			server.ResourceIDToVolumeName("unknown-subsystem-id"),
-			nil,
-			[]string{""},
-			codes.NotFound,
-			fmt.Sprintf("unable to find key %v", server.ResourceIDToVolumeName("unknown-subsystem-id")),
-			false,
-		},
-	}
-
-	// run tests
-	for name, tt := range tests {
-		t.Run(name, func(t *testing.T) {
-			testEnv := createTestEnvironment(tt.start, tt.spdk)
-			defer testEnv.Close()
-
-			testEnv.opiSpdkServer.Subsystems[testSubsystemName] = &testSubsystem
-			testEnv.opiSpdkServer.Controllers[testControllerName] = &testController
-			testEnv.opiSpdkServer.Namespaces[testNamespaceName] = &testNamespace
-
-			request := &pb.GetNvmeControllerRequest{Name: tt.in}
-			response, err := testEnv.client.GetNvmeController(testEnv.ctx, request)
-			if response != nil {
-				if !reflect.DeepEqual(response.Spec, tt.out.Spec) {
-					t.Error("response: expected", tt.out.GetSpec(), "received", response.GetSpec())
-				}
-				if !reflect.DeepEqual(response.Status, tt.out.Status) {
-					t.Error("response: expected", tt.out.GetStatus(), "received", response.GetStatus())
-				}
-			}
-
-			if err != nil {
-				if er, ok := status.FromError(err); ok {
-					if er.Code() != tt.errCode {
-						t.Error("error code: expected", tt.errCode, "received", er.Code())
-					}
-					if er.Message() != tt.errMsg {
-						t.Error("error message: expected", tt.errMsg, "received", er.Message())
-					}
-				}
-			}
-		})
-	}
-}
-
-func TestFrontEnd_NvmeControllerStats(t *testing.T) {
-	tests := map[string]struct {
-		in      string
-		out     *pb.NvmeControllerStatsResponse
-		spdk    []string
-		errCode codes.Code
-		errMsg  string
-		start   bool
-	}{
-		"unimplemented method": {
-			testControllerName,
-			nil,
-			[]string{""},
-			codes.Unimplemented,
-			fmt.Sprintf("%v method is not implemented", "NvmeControllerStats"),
-			false,
-		},
-	}
-
-	// run tests
-	for name, tt := range tests {
-		t.Run(name, func(t *testing.T) {
-			testEnv := createTestEnvironment(tt.start, tt.spdk)
-			defer testEnv.Close()
-
-			testEnv.opiSpdkServer.Subsystems[testSubsystemName] = &testSubsystem
-			testEnv.opiSpdkServer.Controllers[testControllerName] = &testController
-			testEnv.opiSpdkServer.Namespaces[testNamespaceName] = &testNamespace
-
-			request := &pb.NvmeControllerStatsRequest{Id: &pc.ObjectKey{Value: tt.in}}
-			response, err := testEnv.client.NvmeControllerStats(testEnv.ctx, request)
-			if response != nil {
-				t.Error("response: expected", tt.out, "received", response)
-			}
-
-			if err != nil {
-				if er, ok := status.FromError(err); ok {
-					if er.Code() != tt.errCode {
-						t.Error("error code: expected", tt.errCode, "received", er.Code())
-					}
-					if er.Message() != tt.errMsg {
-						t.Error("error message: expected", tt.errMsg, "received", er.Message())
-					}
-				}
-			}
-		})
-	}
-}
-
-func TestFrontEnd_CreateNvmeNamespace(t *testing.T) {
-	spec := &pb.NvmeNamespaceSpec{
-		SubsystemId: &pc.ObjectKey{Value: testSubsystemName},
-		HostNsid:    0,
-		VolumeId:    &pc.ObjectKey{Value: "Malloc1"},
-		Uuid:        &pc.Uuid{Value: "1b4e28ba-2fa1-11d2-883f-b9a761bde3fb"},
-		Nguid:       "1b4e28ba-2fa1-11d2-883f-b9a761bde3fb",
-		Eui64:       1967554867335598546,
-	}
-	namespaceSpec := &pb.NvmeNamespaceSpec{
-		SubsystemId: &pc.ObjectKey{Value: testSubsystemName},
-		HostNsid:    22,
-		VolumeId:    &pc.ObjectKey{Value: "Malloc1"},
-		Uuid:        &pc.Uuid{Value: "1b4e28ba-2fa1-11d2-883f-b9a761bde3fb"},
-		Nguid:       "1b4e28ba-2fa1-11d2-883f-b9a761bde3fb",
-		Eui64:       1967554867335598546,
-	}
-	tests := map[string]struct {
-		id      string
-		in      *pb.NvmeNamespace
-		out     *pb.NvmeNamespace
-		spdk    []string
-		errCode codes.Code
-		errMsg  string
-		start   bool
-		exist   bool
-	}{
-		"illegal resource_id": {
-			"CapitalLettersNotAllowed",
-			&pb.NvmeNamespace{
-				Spec: spec,
-			},
-			nil,
-			[]string{""},
-			codes.Unknown,
-			fmt.Sprintf("user-settable ID must only contain lowercase, numbers and hyphens (%v)", "got: 'C' in position 0"),
-			false,
-			false,
-		},
-		"valid request with invalid SPDK response": {
-			testNamespaceID,
-			&pb.NvmeNamespace{
-				Spec: spec,
-			},
-			nil,
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":false}`},
-			codes.InvalidArgument,
-			fmt.Sprintf("Could not create NS: %v", testNamespaceName),
-			true,
-			false,
-		},
-		"valid request with empty SPDK response": {
-			testNamespaceID,
-			&pb.NvmeNamespace{
-				Spec: spec,
-			},
-			nil,
-			[]string{""},
-			codes.Unknown,
-			fmt.Sprintf("controller_nvme_namespace_attach: %v", "EOF"),
-			true,
-			false,
-		},
-		"valid request with ID mismatch SPDK response": {
-			testNamespaceID,
-			&pb.NvmeNamespace{
-				Spec: spec,
-			},
-			nil,
-			[]string{`{"id":0,"error":{"code":0,"message":""},"result":false}`},
-			codes.Unknown,
-			fmt.Sprintf("controller_nvme_namespace_attach: %v", "json response ID mismatch"),
-			true,
-			false,
-		},
-		"valid request with error code from SPDK response": {
-			testNamespaceID,
-			&pb.NvmeNamespace{
-				Spec: spec,
-			},
-			nil,
-			[]string{`{"id":%d,"error":{"code":1,"message":"myopierr"},"result":false}`},
-			codes.Unknown,
-			fmt.Sprintf("controller_nvme_namespace_attach: %v", "json response error: myopierr"),
-			true,
-			false,
-		},
-		"valid request with valid SPDK response": {
-			testNamespaceID,
-			&pb.NvmeNamespace{
-				Spec: namespaceSpec,
-			},
-			&pb.NvmeNamespace{
-				Name: testNamespaceName,
-				Spec: &pb.NvmeNamespaceSpec{
-					SubsystemId: &pc.ObjectKey{Value: testSubsystemName},
-					HostNsid:    22,
-					VolumeId:    &pc.ObjectKey{Value: "Malloc1"},
-					Uuid:        &pc.Uuid{Value: "1b4e28ba-2fa1-11d2-883f-b9a761bde3fb"},
-					Nguid:       "1b4e28ba-2fa1-11d2-883f-b9a761bde3fb",
-					Eui64:       1967554867335598546,
-				},
-				Status: &pb.NvmeNamespaceStatus{
-					PciState:     2,
-					PciOperState: 1,
-				},
-			},
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":true}`},
-			codes.OK,
-			"",
-			true,
-			false,
-		},
-		"already exists": {
-			testNamespaceID,
-			&pb.NvmeNamespace{
-				Spec: spec,
-			},
-			&testNamespace,
-			[]string{""},
-			codes.OK,
-			"",
-			false,
-			true,
-		},
-	}
-
-	// run tests
-	for name, tt := range tests {
-		t.Run(name, func(t *testing.T) {
-			testEnv := createTestEnvironment(tt.start, tt.spdk)
-			defer testEnv.Close()
-
-			testEnv.opiSpdkServer.Subsystems[testSubsystemName] = &testSubsystem
-			testEnv.opiSpdkServer.Controllers[testControllerName] = &testController
-			if tt.exist {
-				testEnv.opiSpdkServer.Namespaces[testNamespaceName] = &testNamespace
-				testEnv.opiSpdkServer.Namespaces[testNamespaceName].Name = testNamespaceName
-			}
-			if tt.out != nil {
-				tt.out.Name = testNamespaceName
-			}
-
-			request := &pb.CreateNvmeNamespaceRequest{NvmeNamespace: tt.in, NvmeNamespaceId: tt.id}
-			response, err := testEnv.client.CreateNvmeNamespace(testEnv.ctx, request)
-			if response != nil {
-				mtt, _ := proto.Marshal(tt.out)
-				mResponse, _ := proto.Marshal(response)
-				if !bytes.Equal(mtt, mResponse) {
-					t.Error("response: expected", tt.out, "received", response)
-				}
-			}
-
-			if err != nil {
-				if er, ok := status.FromError(err); ok {
-					if er.Code() != tt.errCode {
-						t.Error("error code: expected", tt.errCode, "received", er.Code())
-					}
-					if er.Message() != tt.errMsg {
-						t.Error("error message: expected", tt.errMsg, "received", er.Message())
-					}
-				}
-			}
-		})
-	}
-}
-
-func TestFrontEnd_UpdateNvmeNamespace(t *testing.T) {
-	tests := map[string]struct {
-		mask    *fieldmaskpb.FieldMask
-		in      *pb.NvmeNamespace
-		out     *pb.NvmeNamespace
-		spdk    []string
-		errCode codes.Code
-		errMsg  string
-		start   bool
-	}{
-		"invalid fieldmask": {
-			&fieldmaskpb.FieldMask{Paths: []string{"*", "author"}},
-			&pb.NvmeNamespace{
-				Name: testNamespaceName,
-			},
-			nil,
-			[]string{""},
-			codes.Unknown,
-			fmt.Sprintf("invalid field path: %s", "'*' must not be used with other paths"),
-			false,
-		},
-		"unimplemented method": {
-			nil,
-			&pb.NvmeNamespace{
-				Name: testNamespaceName,
-			},
-			nil,
-			[]string{""},
-			codes.Unimplemented,
-			fmt.Sprintf("%v method is not implemented", "UpdateNvmeNamespace"),
-			false,
-		},
-		"valid request with unknown key": {
-			nil,
-			&pb.NvmeNamespace{
-				Name: server.ResourceIDToVolumeName("unknown-id"),
-			},
-			nil,
-			[]string{""},
-			codes.NotFound,
-			fmt.Sprintf("unable to find key %v", server.ResourceIDToVolumeName("unknown-id")),
-			false,
-		},
-	}
-
-	// run tests
-	for name, tt := range tests {
-		t.Run(name, func(t *testing.T) {
-			testEnv := createTestEnvironment(tt.start, tt.spdk)
-			defer testEnv.Close()
-
-			testEnv.opiSpdkServer.Subsystems[testSubsystemName] = &testSubsystem
-			testEnv.opiSpdkServer.Controllers[testControllerName] = &testController
-			testEnv.opiSpdkServer.Namespaces[testNamespaceName] = &testNamespace
-
-			request := &pb.UpdateNvmeNamespaceRequest{NvmeNamespace: tt.in, UpdateMask: tt.mask}
-			response, err := testEnv.client.UpdateNvmeNamespace(testEnv.ctx, request)
-			if response != nil {
-				t.Error("response: expected", tt.out, "received", response)
-			}
-
-			if err != nil {
-				if er, ok := status.FromError(err); ok {
-					if er.Code() != tt.errCode {
-						t.Error("error code: expected", tt.errCode, "received", er.Code())
-					}
-					if er.Message() != tt.errMsg {
-						t.Error("error message: expected", tt.errMsg, "received", er.Message())
-					}
-				}
-			}
-		})
-	}
-}
-
-func TestFrontEnd_ListNvmeNamespaces(t *testing.T) {
-	tests := map[string]struct {
-		in      string
-		out     []*pb.NvmeNamespace
-		spdk    []string
-		errCode codes.Code
-		errMsg  string
-		start   bool
-		size    int32
-		token   string
-	}{
-		"valid request with invalid SPDK response": {
-			testSubsystemName,
-			nil,
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":{"name":"","cntlid":0,"Namespaces":null}}`},
-			codes.InvalidArgument,
-			fmt.Sprintf("Could not create NQN: %v", "nqn.2022-09.io.spdk:opi3"),
-			true,
-			0,
-			"",
-		},
-		"valid request with invalid marshal SPDK response": {
-			testSubsystemName,
-			nil,
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":[]}`},
-			codes.Unknown,
-			fmt.Sprintf("controller_nvme_namespace_list: %v", "json: cannot unmarshal array into Go value of type models.NvdaControllerNvmeNamespaceListResult"),
-			true,
-			0,
-			"",
-		},
-		"valid request with empty SPDK response": {
-			testSubsystemName,
-			nil,
-			[]string{""},
-			codes.Unknown,
-			fmt.Sprintf("controller_nvme_namespace_list: %v", "EOF"),
-			true,
-			0,
-			"",
-		},
-		"valid request with ID mismatch SPDK response": {
-			testSubsystemName,
-			nil,
-			[]string{`{"id":0,"error":{"code":0,"message":""},"result":{"name":"","cntlid":0,"Namespaces":null}}`},
-			codes.Unknown,
-			fmt.Sprintf("controller_nvme_namespace_list: %v", "json response ID mismatch"),
-			true,
-			0,
-			"",
-		},
-		"valid request with error code from SPDK response": {
-			testSubsystemName,
-			nil,
-			[]string{`{"id":%d,"error":{"code":1,"message":"myopierr"}}`},
-			codes.Unknown,
-			fmt.Sprintf("controller_nvme_namespace_list: %v", "json response error: myopierr"),
-			true,
-			0,
-			"",
-		},
-		"pagination negative": {
-			testSubsystemName,
-			nil,
-			[]string{},
-			codes.InvalidArgument,
-			"negative PageSize is not allowed",
-			false,
-			-10,
-			"",
-		},
-		"pagination error": {
-			testSubsystemName,
-			nil,
-			[]string{},
-			codes.NotFound,
-			fmt.Sprintf("unable to find pagination token %s", "unknown-pagination-token"),
-			false,
-			0,
-			"unknown-pagination-token",
-		},
-		"pagination": {
-			testSubsystemName,
-			[]*pb.NvmeNamespace{
-				{
-					Spec: &pb.NvmeNamespaceSpec{
-						HostNsid: 11,
-					},
-				},
-			},
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":{"name": "NvmeEmu0pf1", "cntlid": 1, "Namespaces": [{"nsid": 11, "bdev": "Malloc0", "bdev_type": "spdk", "qn": "", "protocol": ""},{"nsid": 12, "bdev": "Malloc1", "bdev_type": "spdk", "qn": "", "protocol": ""},{"nsid": 13, "bdev": "Malloc2", "bdev_type": "spdk", "qn": "", "protocol": ""}]}}`},
-			codes.OK,
-			"",
-			true,
-			1,
-			"",
-		},
-		"pagination overflow": {
-			testSubsystemName,
-			[]*pb.NvmeNamespace{
-				{
-					Spec: &pb.NvmeNamespaceSpec{
-						HostNsid: 11,
-					},
-				},
-				{
-					Spec: &pb.NvmeNamespaceSpec{
-						HostNsid: 12,
-					},
-				},
-				{
-					Spec: &pb.NvmeNamespaceSpec{
-						HostNsid: 13,
-					},
-				},
-			},
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":{"name": "NvmeEmu0pf1", "cntlid": 1, "Namespaces": [{"nsid": 11, "bdev": "Malloc0", "bdev_type": "spdk", "qn": "", "protocol": ""},{"nsid": 12, "bdev": "Malloc1", "bdev_type": "spdk", "qn": "", "protocol": ""},{"nsid": 13, "bdev": "Malloc2", "bdev_type": "spdk", "qn": "", "protocol": ""}]}}`},
-			codes.OK,
-			"",
-			true,
-			1000,
-			"",
-		},
-		"pagination offset": {
-			testSubsystemName,
-			[]*pb.NvmeNamespace{
-				{
-					Spec: &pb.NvmeNamespaceSpec{
-						HostNsid: 12,
-					},
-				},
-			},
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":{"name": "NvmeEmu0pf1", "cntlid": 1, "Namespaces": [{"nsid": 11, "bdev": "Malloc0", "bdev_type": "spdk", "qn": "", "protocol": ""},{"nsid": 12, "bdev": "Malloc1", "bdev_type": "spdk", "qn": "", "protocol": ""},{"nsid": 13, "bdev": "Malloc2", "bdev_type": "spdk", "qn": "", "protocol": ""}]}}`},
-			codes.OK,
-			"",
-			true,
-			1,
-			"existing-pagination-token",
-		},
-		"valid request with valid SPDK response": {
-			testSubsystemName,
-			[]*pb.NvmeNamespace{
-				{
-					Spec: &pb.NvmeNamespaceSpec{
-						HostNsid: 11,
-					},
-				},
-				{
-					Spec: &pb.NvmeNamespaceSpec{
-						HostNsid: 12,
-					},
-				},
-				{
-					Spec: &pb.NvmeNamespaceSpec{
-						HostNsid: 13,
-					},
-				},
-			},
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":{"name": "NvmeEmu0pf1", "cntlid": 1, "Namespaces": [{"nsid": 11, "bdev": "Malloc0", "bdev_type": "spdk", "qn": "", "protocol": ""},{"nsid": 12, "bdev": "Malloc1", "bdev_type": "spdk", "qn": "", "protocol": ""},{"nsid": 13, "bdev": "Malloc2", "bdev_type": "spdk", "qn": "", "protocol": ""}]}}`},
-			codes.OK,
-			"",
-			true,
-			0,
-			"",
-		},
-		"valid request with unknown key": {
-			server.ResourceIDToVolumeName("unknown-namespace-id"),
-			nil,
-			[]string{""},
-			codes.NotFound,
-			fmt.Sprintf("unable to find key %v", server.ResourceIDToVolumeName("unknown-namespace-id")),
-			false,
-			0,
-			"",
-		},
-	}
-
-	// run tests
-	for name, tt := range tests {
-		t.Run(name, func(t *testing.T) {
-			testEnv := createTestEnvironment(tt.start, tt.spdk)
-			defer testEnv.Close()
-
-			testEnv.opiSpdkServer.Subsystems[testSubsystemName] = &testSubsystem
-			testEnv.opiSpdkServer.Controllers[testControllerName] = &testController
-			testEnv.opiSpdkServer.Namespaces[testNamespaceName] = &testNamespace
-			testEnv.opiSpdkServer.Pagination["existing-pagination-token"] = 1
-
-			request := &pb.ListNvmeNamespacesRequest{Parent: tt.in, PageSize: tt.size, PageToken: tt.token}
-			response, err := testEnv.client.ListNvmeNamespaces(testEnv.ctx, request)
-			if response != nil {
-				if !reflect.DeepEqual(response.NvmeNamespaces, tt.out) {
-					t.Error("response: expected", tt.out, "received", response.NvmeNamespaces)
-				}
-			}
-
-			if err != nil {
-				if er, ok := status.FromError(err); ok {
-					if er.Code() != tt.errCode {
-						t.Error("error code: expected", tt.errCode, "received", er.Code())
-					}
-					if er.Message() != tt.errMsg {
-						t.Error("error message: expected", tt.errMsg, "received", er.Message())
-					}
-				}
-			}
-		})
-	}
-}
-
-func TestFrontEnd_GetNvmeNamespace(t *testing.T) {
-	tests := map[string]struct {
-		in      string
-		out     *pb.NvmeNamespace
-		spdk    []string
-		errCode codes.Code
-		errMsg  string
-		start   bool
-	}{
-		"valid request with invalid SPDK response": {
-			testNamespaceName,
-			nil,
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":{"name":"","cntlid":17,"Namespaces":null}}`},
-			codes.InvalidArgument,
-			fmt.Sprintf("Could not find HostNsid: %v", "22"),
-			true,
-		},
-		"valid request with invalid marshal SPDK response": {
-			testNamespaceName,
-			nil,
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":[]}`},
-			codes.Unknown,
-			fmt.Sprintf("controller_nvme_namespace_list: %v", "json: cannot unmarshal array into Go value of type models.NvdaControllerNvmeNamespaceListResult"),
-			true,
-		},
-		"valid request with empty SPDK response": {
-			testNamespaceName,
-			nil,
-			[]string{""},
-			codes.Unknown,
-			fmt.Sprintf("controller_nvme_namespace_list: %v", "EOF"),
-			true,
-		},
-		"valid request with ID mismatch SPDK response": {
-			testNamespaceName,
-			nil,
-			[]string{`{"id":0,"error":{"code":0,"message":""},"result":{"name":"","cntlid":0,"Namespaces":null}}`},
-			codes.Unknown,
-			fmt.Sprintf("controller_nvme_namespace_list: %v", "json response ID mismatch"),
-			true,
-		},
-		"valid request with error code from SPDK response": {
-			testNamespaceName,
-			nil,
-			[]string{`{"id":%d,"error":{"code":1,"message":"myopierr"}}`},
-			codes.Unknown,
-			fmt.Sprintf("controller_nvme_namespace_list: %v", "json response error: myopierr"),
-			true,
-		},
-		"valid request with valid SPDK response": {
-			testNamespaceName,
-			&pb.NvmeNamespace{
-				Spec: &pb.NvmeNamespaceSpec{
-					HostNsid: 22,
-				},
-				Status: &pb.NvmeNamespaceStatus{
-					PciState:     2,
-					PciOperState: 1,
-				},
-			},
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":{"name": "NvmeEmu0pf1", "cntlid": 1, "Namespaces": [{"nsid": 11, "bdev": "Malloc0", "bdev_type": "spdk", "qn": "", "protocol": ""},{"nsid": 22, "bdev": "Malloc1", "bdev_type": "spdk", "qn": "", "protocol": ""},{"nsid": 13, "bdev": "Malloc2", "bdev_type": "spdk", "qn": "", "protocol": ""}]}}`},
-			codes.OK,
-			"",
-			true,
-		},
-		"valid request with unknown key": {
-			server.ResourceIDToVolumeName("unknown-namespace-id"),
-			nil,
-			[]string{""},
-			codes.NotFound,
-			fmt.Sprintf("unable to find key %v", server.ResourceIDToVolumeName("unknown-namespace-id")),
-			false,
-		},
-	}
-
-	// run tests
-	for name, tt := range tests {
-		t.Run(name, func(t *testing.T) {
-			testEnv := createTestEnvironment(tt.start, tt.spdk)
-			defer testEnv.Close()
-
-			testEnv.opiSpdkServer.Subsystems[testSubsystemName] = &testSubsystem
-			testEnv.opiSpdkServer.Controllers[testControllerName] = &testController
-			testEnv.opiSpdkServer.Namespaces[testNamespaceName] = &testNamespace
-
-			request := &pb.GetNvmeNamespaceRequest{Name: tt.in}
-			response, err := testEnv.client.GetNvmeNamespace(testEnv.ctx, request)
-			if response != nil {
-				if !reflect.DeepEqual(response.Spec, tt.out.Spec) {
-					t.Error("response: expected", tt.out.GetSpec(), "received", response.GetSpec())
-				}
-				if !reflect.DeepEqual(response.Status, tt.out.Status) {
-					t.Error("response: expected", tt.out.GetStatus(), "received", response.GetStatus())
-				}
-			}
-
-			if err != nil {
-				if er, ok := status.FromError(err); ok {
-					if er.Code() != tt.errCode {
-						t.Error("error code: expected", tt.errCode, "received", er.Code())
-					}
-					if er.Message() != tt.errMsg {
-						t.Error("error message: expected", tt.errMsg, "received", er.Message())
-					}
-				}
-			}
-		})
-	}
-}
-
-func TestFrontEnd_NvmeNamespaceStats(t *testing.T) {
-	tests := map[string]struct {
-		in      string
-		out     *pb.VolumeStats
-		spdk    []string
-		errCode codes.Code
-		errMsg  string
-		start   bool
-	}{
-		"valid request with invalid SPDK response": {
-			testNamespaceName,
-			nil,
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":{"controllers":[{"name":"NvmeEmu0pf1","bdevs":[]}]}}`},
-			codes.InvalidArgument,
-			fmt.Sprintf("Could not find BdevName: %v", "Malloc1"),
-			true,
-		},
-		"valid request with invalid marshal SPDK response": {
-			testNamespaceName,
-			nil,
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":[]}`},
-			codes.Unknown,
-			fmt.Sprintf("controller_nvme_get_iostat: %v", "json: cannot unmarshal array into Go value of type models.NvdaControllerNvmeStatsResult"),
-			true,
-		},
-		"valid request with empty SPDK response": {
-			testNamespaceName,
-			nil,
-			[]string{""},
-			codes.Unknown,
-			fmt.Sprintf("controller_nvme_get_iostat: %v", "EOF"),
-			true,
-		},
-		"valid request with ID mismatch SPDK response": {
-			testNamespaceName,
-			nil,
-			[]string{`{"id":0,"error":{"code":0,"message":""},"result":{"controllers":[{"name":"NvmeEmu0pf1","bdevs":[]}]}}`},
-			codes.Unknown,
-			fmt.Sprintf("controller_nvme_get_iostat: %v", "json response ID mismatch"),
-			true,
-		},
-		"valid request with error code from SPDK response": {
-			testNamespaceName,
-			nil,
-			[]string{`{"id":%d,"error":{"code":1,"message":"myopierr"}}`},
-			codes.Unknown,
-			fmt.Sprintf("controller_nvme_get_iostat: %v", "json response error: myopierr"),
-			true,
-		},
-		"valid request with valid SPDK response": {
-			testNamespaceName,
-			&pb.VolumeStats{
-				ReadOpsCount:  12345,
-				WriteOpsCount: 54321,
-			},
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result": {"controllers":[{"name":"NvmeEmu0pf1","bdevs":[{"bdev_name":"Malloc0","read_ios":55,"completed_read_ios":55,"write_ios":33,"completed_write_ios":33,"flush_ios":0,"completed_flush_ios":0,"err_read_ios":0,"err_write_ios":0,"err_flush_ios":0},{"bdev_name":"Malloc1","read_ios":12345,"completed_read_ios":12345,"write_ios":54321,"completed_write_ios":54321,"flush_ios":0,"completed_flush_ios":0,"err_read_ios":0,"err_write_ios":0,"err_flush_ios":0}]}]}}`},
-			codes.OK,
-			"",
-			true,
-		},
-		"valid request with unknown key": {
-			"unknown-namespace-id",
-			nil,
-			[]string{""},
-			codes.NotFound,
-			fmt.Sprintf("unable to find key %v", "unknown-namespace-id"),
-			false,
-		},
-	}
-
-	// run tests
-	for name, tt := range tests {
-		t.Run(name, func(t *testing.T) {
-			testEnv := createTestEnvironment(tt.start, tt.spdk)
-			defer testEnv.Close()
-
-			testEnv.opiSpdkServer.Subsystems[testSubsystemName] = &testSubsystem
-			testEnv.opiSpdkServer.Controllers[testControllerName] = &testController
-			testEnv.opiSpdkServer.Namespaces[testNamespaceName] = &testNamespace
-
-			request := &pb.NvmeNamespaceStatsRequest{NamespaceId: &pc.ObjectKey{Value: tt.in}}
-			response, err := testEnv.client.NvmeNamespaceStats(testEnv.ctx, request)
-			if response != nil {
-				if !reflect.DeepEqual(response.Stats, tt.out) {
-					t.Error("response: expected", tt.out, "received", response.Stats)
-				}
-			}
-
-			if err != nil {
-				if er, ok := status.FromError(err); ok {
-					if er.Code() != tt.errCode {
-						t.Error("error code: expected", tt.errCode, "received", er.Code())
-					}
-					if er.Message() != tt.errMsg {
-						t.Error("error message: expected", tt.errMsg, "received", er.Message())
-					}
-				}
-			}
-		})
-	}
-}
-
-func TestFrontEnd_DeleteNvmeNamespace(t *testing.T) {
-	tests := map[string]struct {
-		in      string
-		out     *emptypb.Empty
-		spdk    []string
-		errCode codes.Code
-		errMsg  string
-		start   bool
-		missing bool
-	}{
-		"valid request with invalid SPDK response": {
-			testNamespaceName,
-			nil,
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":false}`},
-			codes.InvalidArgument,
-			fmt.Sprintf("Could not delete NS: %v", testSubsystemName),
-			true,
-			false,
-		},
-		"valid request with empty SPDK response": {
-			testNamespaceName,
-			nil,
-			[]string{""},
-			codes.Unknown,
-			fmt.Sprintf("controller_nvme_namespace_detach: %v", "EOF"),
-			true,
-			false,
-		},
-		"valid request with ID mismatch SPDK response": {
-			testNamespaceName,
-			nil,
-			[]string{`{"id":0,"error":{"code":0,"message":""},"result":false}`},
-			codes.Unknown,
-			fmt.Sprintf("controller_nvme_namespace_detach: %v", "json response ID mismatch"),
-			true,
-			false,
-		},
-		"valid request with error code from SPDK response": {
-			testNamespaceName,
-			nil,
-			[]string{`{"id":%d,"error":{"code":1,"message":"myopierr"},"result":false}`},
-			codes.Unknown,
-			fmt.Sprintf("controller_nvme_namespace_detach: %v", "json response error: myopierr"),
-			true,
-			false,
-		},
-		"valid request with valid SPDK response": {
-			testNamespaceName,
-			&emptypb.Empty{},
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":true}`}, // `{"jsonrpc": "2.0", "id": 1, "result": True}`,
-			codes.OK,
-			"",
-			true,
-			false,
-		},
-		"valid request with unknown key": {
-			"unknown-namespace-id",
-			nil,
-			[]string{""},
-			codes.NotFound,
-			fmt.Sprintf("unable to find key %v", "unknown-namespace-id"),
-			false,
-			false,
-		},
-		"unknown key with missing allowed": {
-			"unknown-id",
-			&emptypb.Empty{},
-			[]string{""},
-			codes.OK,
-			"",
-			false,
-			true,
-		},
-	}
-
-	// run tests
-	for name, tt := range tests {
-		t.Run(name, func(t *testing.T) {
-			testEnv := createTestEnvironment(tt.start, tt.spdk)
-			defer testEnv.Close()
-
-			testEnv.opiSpdkServer.Subsystems[testSubsystemName] = &testSubsystem
-			testEnv.opiSpdkServer.Controllers[testControllerName] = &testController
-			testEnv.opiSpdkServer.Namespaces[testNamespaceName] = &testNamespace
-
-			request := &pb.DeleteNvmeNamespaceRequest{Name: tt.in, AllowMissing: tt.missing}
-			response, err := testEnv.client.DeleteNvmeNamespace(testEnv.ctx, request)
-			if err != nil {
-				if er, ok := status.FromError(err); ok {
-					if er.Code() != tt.errCode {
-						t.Error("error code: expected", tt.errCode, "received", er.Code())
-					}
-					if er.Message() != tt.errMsg {
-						t.Error("error message: expected", tt.errMsg, "received", er.Message())
-					}
-				}
-			}
-			if reflect.TypeOf(response) != reflect.TypeOf(tt.out) {
-				t.Error("response: expected", reflect.TypeOf(tt.out), "received", reflect.TypeOf(response))
-			}
-		})
-	}
-}
-
-func TestFrontEnd_DeleteNvmeController(t *testing.T) {
-	tests := map[string]struct {
-		in      string
-		out     *emptypb.Empty
-		spdk    []string
-		errCode codes.Code
-		errMsg  string
-		start   bool
-		missing bool
-	}{
-		"valid request with invalid SPDK response": {
-			testControllerName,
-			nil,
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":false}`},
-			codes.InvalidArgument,
-			fmt.Sprintf("Could not delete NQN:ID %v", "nqn.2022-09.io.spdk:opi3:17"),
-			true,
-			false,
-		},
-		"valid request with empty SPDK response": {
-			testControllerName,
-			nil,
-			[]string{""},
-			codes.Unknown,
-			fmt.Sprintf("controller_nvme_delete: %v", "EOF"),
-			true,
-			false,
-		},
-		"valid request with ID mismatch SPDK response": {
-			testControllerName,
-			nil,
-			[]string{`{"id":0,"error":{"code":0,"message":""},"result":false}`},
-			codes.Unknown,
-			fmt.Sprintf("controller_nvme_delete: %v", "json response ID mismatch"),
-			true,
-			false,
-		},
-		"valid request with error code from SPDK response": {
-			testControllerName,
-			nil,
-			[]string{`{"id":%d,"error":{"code":1,"message":"myopierr"},"result":false}`},
-			codes.Unknown,
-			fmt.Sprintf("controller_nvme_delete: %v", "json response error: myopierr"),
-			true,
-			false,
-		},
-		"valid request with valid SPDK response": {
-			testControllerName,
-			&emptypb.Empty{},
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":true}`}, // `{"jsonrpc": "2.0", "id": 1, "result": True}`,
-			codes.OK,
-			"",
-			true,
-			false,
-		},
-		"valid request with unknown key": {
-			"unknown-controller-id",
-			nil,
-			[]string{""},
-			codes.Unknown,
-			fmt.Sprintf("error finding controller %v", "unknown-controller-id"),
-			false,
-			false,
-		},
-		"unknown key with missing allowed": {
-			"unknown-id",
-			&emptypb.Empty{},
-			[]string{""},
-			codes.OK,
-			"",
-			false,
-			true,
-		},
-	}
-
-	// run tests
-	for name, tt := range tests {
-		t.Run(name, func(t *testing.T) {
-			testEnv := createTestEnvironment(tt.start, tt.spdk)
-			defer testEnv.Close()
-
-			testEnv.opiSpdkServer.Subsystems[testSubsystemName] = &testSubsystem
-			testEnv.opiSpdkServer.Controllers[testControllerName] = &testController
-			testEnv.opiSpdkServer.Namespaces[testNamespaceName] = &testNamespace
-
-			request := &pb.DeleteNvmeControllerRequest{Name: tt.in, AllowMissing: tt.missing}
-			response, err := testEnv.client.DeleteNvmeController(testEnv.ctx, request)
-			if err != nil {
-				if er, ok := status.FromError(err); ok {
-					if er.Code() != tt.errCode {
-						t.Error("error code: expected", tt.errCode, "received", er.Code())
-					}
-					if er.Message() != tt.errMsg {
-						t.Error("error message: expected", tt.errMsg, "received", er.Message())
-					}
-				}
-			}
-			if reflect.TypeOf(response) != reflect.TypeOf(tt.out) {
-				t.Error("response: expected", reflect.TypeOf(tt.out), "received", reflect.TypeOf(response))
-			}
-		})
-	}
-}
-
-func TestFrontEnd_DeleteNvmeSubsystem(t *testing.T) {
-	tests := map[string]struct {
-		in      string
-		out     *emptypb.Empty
-		spdk    []string
-		errCode codes.Code
-		errMsg  string
-		start   bool
-		missing bool
-	}{
-		"valid request with invalid SPDK response": {
-			testSubsystemName,
-			nil,
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":false}`},
-			codes.InvalidArgument,
-			fmt.Sprintf("Could not delete NQN: %v", "nqn.2022-09.io.spdk:opi3"),
-			true,
-			false,
-		},
-		"valid request with empty SPDK response": {
-			testSubsystemName,
-			nil,
-			[]string{""},
-			codes.Unknown,
-			fmt.Sprintf("subsystem_nvme_delete: %v", "EOF"),
-			true,
-			false,
-		},
-		"valid request with ID mismatch SPDK response": {
-			testSubsystemName,
-			nil,
-			[]string{`{"id":0,"error":{"code":0,"message":""},"result":false}`},
-			codes.Unknown,
-			fmt.Sprintf("subsystem_nvme_delete: %v", "json response ID mismatch"),
-			true,
-			false,
-		},
-		"valid request with error code from SPDK response": {
-			testSubsystemName,
-			nil,
-			[]string{`{"id":%d,"error":{"code":1,"message":"myopierr"},"result":false}`},
-			codes.Unknown,
-			fmt.Sprintf("subsystem_nvme_delete: %v", "json response error: myopierr"),
-			true,
-			false,
-		},
-		"valid request with valid SPDK response": {
-			testSubsystemName,
-			&emptypb.Empty{},
-			[]string{`{"id":%d,"error":{"code":0,"message":""},"result":true}`}, // `{"jsonrpc": "2.0", "id": 1, "result": True}`,
-			codes.OK,
-			"",
-			true,
-			false,
-		},
-		"valid request with unknown key": {
-			"unknown-subsystem-id",
-			nil,
-			[]string{""},
-			codes.NotFound,
-			fmt.Sprintf("unable to find key %v", "unknown-subsystem-id"),
-			false,
-			false,
-		},
-		"unknown key with missing allowed": {
-			"unknown-id",
-			&emptypb.Empty{},
-			[]string{""},
-			codes.OK,
-			"",
-			false,
-			true,
-		},
-	}
-
-	// run tests
-	for name, tt := range tests {
-		t.Run(name, func(t *testing.T) {
-			testEnv := createTestEnvironment(tt.start, tt.spdk)
-			defer testEnv.Close()
-
-			testEnv.opiSpdkServer.Subsystems[testSubsystemName] = &testSubsystem
-			testEnv.opiSpdkServer.Controllers[testControllerName] = &testController
-			testEnv.opiSpdkServer.Namespaces[testNamespaceName] = &testNamespace
-
-			request := &pb.DeleteNvmeSubsystemRequest{Name: tt.in, AllowMissing: tt.missing}
-			response, err := testEnv.client.DeleteNvmeSubsystem(testEnv.ctx, request)
-			if err != nil {
-				if er, ok := status.FromError(err); ok {
-					if er.Code() != tt.errCode {
-						t.Error("error code: expected", tt.errCode, "received", er.Code())
-					}
-					if er.Message() != tt.errMsg {
-						t.Error("error message: expected", tt.errMsg, "received", er.Message())
-					}
-				}
-			}
-			if reflect.TypeOf(response) != reflect.TypeOf(tt.out) {
-				t.Error("response: expected", reflect.TypeOf(tt.out), "received", reflect.TypeOf(response))
-			}
-		})
-	}
-}
